% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ao_localRelief.R
\name{ao_pc2Hillshade}
\alias{ao_pc2Hillshade}
\title{Create a hillshading model l from a UAV generated point cloud}
\usage{
ao_pc2Hillshade(lasDir = NULL, gisdbase_path = NULL,
  GRASSlocation = "tmp/", projFolder = c("data/", "output/", "run/",
  "las/"), grid_size = "1.0", param_list = c("i", "v", "n", "g", "f",
  "overwrite", "quiet"), spline_level_max = "9",
  proj4 = "+proj=utm +zone=32 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs",
  gisdbase_exist = FALSE, path_lastools = NULL)
}
\arguments{
\item{lasDir}{default is \code{NULL} path  to the laz/las file(s)}

\item{gisdbase_path}{gisdbase will be linked or created depending on \code{gisdbase_exist}}

\item{GRASSlocation}{location will be linked or created depending on \code{gisdbase_exist}}

\item{projFolder}{subfolders in gisdbase for R related processing}

\item{grid_size}{resolution for raster operations}

\item{param_list}{default is c("i","v", "n", "g","f","overwrite","quiet")
i Save intermediate maps; \cr
v Use bspline interpolation to construct the surface Uses v.surf.bspline cubic interpolation instead of r.fillnulls cubic interpolation\cr 
n Invert colors in the color table \cr
g Logarithmic scaling of the color table\cr
f Do not perform histogram equalization on the color table}

\item{spline_level_max}{default is 9 number ob spline iterations}

\item{proj4}{default is EPSG 32632 any valid proj4 string that is assumingly the correct one}

\item{gisdbase_exist}{switch if gisdbase is created or  linked only}

\item{path_lastools}{directory for the windows lastools}
}
\description{
Create a hillshading model l from a UAV generated point cloud. return ao_pc2Hillshade basically returns a  DSM
}
\examples{
\dontrun{
# create a hillshade based on a las/laz point clouds 
  hs <- uavRst::ao_pc2Hillshade(lasDir =  "~/proj/Monte_Bernorio/las/",
                             gisdbase_path = "~/temp55/GRASS7",
                             grid_size = "1.0")
}

}
\author{
Chris Reudenbach
}
