% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foa_crownMetrics.R
\name{caMetrics}
\alias{caMetrics}
\title{calculate morphometric features of polygons}
\usage{
caMetrics(crownarea, funNames = c("area", "length", "elongation",
  "eccentricityboundingbox", "solidity", "eccentricityeigen", "calliper",
  "rectangularity", "circularityharalick", "convexity"))
}
