% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uavRst-package.R
\docType{data}
\name{chm_seg}
\alias{chm_seg}
\title{Canopy height model raster map}
\format{\code{"raster::raster"}}
\description{
Example data set containing the canopy height map of a small plot sampled in the Maburg University Forest (MOF). The resolution is 10 cm. ETRS89 UTM32
}
