% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_crownseg.R
\name{chmseg_FT}
\alias{chmseg_FT}
\title{Watershed segmentation based on 'ForestTools'}
\usage{
chmseg_FT(treepos = NULL, chm = NULL, minTreeAlt = 2,
  format = "polygons", verbose = FALSE)
}
\arguments{
\item{treepos}{\code{SpatialPointsDataFrame}. The point locations of treetops. The function will generally produce a
number of crown segments equal to the number of treetops.}

\item{chm}{raster*. Canopy height model in \code{raster} format. Should be the same that was used to create
the input for \code{treepos}.}

\item{minTreeAlt}{numeric. The minimum height value for a \code{CHM} pixel to be considered as part of a crown segment.
All \code{chm} pixels beneath this value will be masked out. Note that this value should be lower than the minimum
height of \code{treepos}.}

\item{format}{character. Format of the function's output. Can be set to either 'raster' or 'polygons'.}

\item{verbose}{to be quiet FALSE}
}
\description{
'ForestTools' segmentation of individual tree crowns based on a canopy height model and initial seeding points (trees). Very fast algorithm based on the imagr watershed algorithm.
Andrew Plowright: R package \href{https://CRAN.R-project.org/package=ForestTools}{'ForestTools'}
}
\examples{
\dontrun{

# required packages
require(uavRst)
require(curl)

# check if SAGA is correctly installed 
if (length(link2GI::findSAGA()) < 1) stop("No valid SAGA GIS instalation found")

# project folder
projRootDir<-tempdir()

# create subfolders please mind that the pathes are exported as global variables
 paths<-link2GI::initProj(projRootDir = projRootDir,
                         projFolders = c("data/","data/ref/","output/","run/","las/"),
                         global = TRUE,
                         path_prefix = "path_")
## overide trailing backslash issue
 path_run<-ifelse(Sys.info()["sysname"]=="Windows", sub("/$", "",path_run),path_run)
 setwd(path_run)                                          
 unlink(paste0(path_run,"*"), force = TRUE
                 
# get the data
 url <- "https://github.com/gisma/gismaData/raw/master/uavRst/data/tutorial.zip"
 res <- curl::curl_download(url, paste0(path_run,"tutorial.zip"))
 unzip(zipfile = res, exdir = path_run)

# make the folders and linkages
 giLinks<-uavRst::get_gi()

# read chm data
 chmR<- raster::raster(paste0(path_run,"chm_2.tif"))
 tPos<- raster::raster(paste0(path_run,"treepos_2.tif"))

# segmentation
 crownsFT <- chmseg_FT(chm = chmR,
                       treepos = tPos,
                       format = "polygons",
                       minTreeAlt = 2,
                       verbose = FALSE)
}
}
