% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_crownSegmentation.R
\name{chmSegmentation}
\alias{chmSegmentation}
\title{seeded region growing tree crown segmentation based on 'SAGA GIS'}
\usage{
chmSegmentation(treePos = NULL, chm = NULL, minTreeAlt = 2,
  minTreeAltParam = "chmQ20", leafsize = 256, normalize = 0,
  neighbour = 1, method = 0, thVarFeature = 1, thVarSpatial = 1,
  thSimilarity = 0.002, seed_params = c("chm"), majority_radius = 3,
  giLinks = NULL)
}
\arguments{
\item{treePos}{spatial raster object}

\item{chm}{Canopy height model in \link[raster]{raster} format. Should be the same that was used to create
the input for \code{treePos}.}

\item{minTreeAlt}{numeric. The minimum height value for a \code{chm} pixel to be considered as part of a crown segment.
All \code{chm} pixels beneath this value will be masked out. Note that this value should be lower than the minimum
height of \code{treePos}.}

\item{minTreeAltParam}{default is "chmQ20"}

\item{leafsize}{integer. bin size of grey value sampling range from 1 to 256}

\item{normalize}{integer.  logical switch if data will be normalized (1)}

\item{neighbour}{integer.  von Neumanns' neighborhood (0) or Moore's (1)}

\item{method}{integer. growing algorithm for feature space and position (0) or feature space only (1)}

\item{thVarFeature}{numeric. spatial variance}

\item{thVarSpatial}{numeric. spatial variance}

\item{thSimilarity}{mumeric. similarity threshold}

\item{seed_params}{vector. of characters corresponding with the used attributes. The altitude values from surface model \code{c("chm")} is mandantory.}

\item{giLinks}{list. of GI tools cli pathes}
}
\description{
Tree segmentation based on a CHM, basically returns a  vector data sets with the tree crown geometries and a bunch of corresponding indices. After the segementation itself the results are hole filled and optionally filtered by a majority filter in the 3*3 surrounding.
}
\examples{
\dontrun{
# crown segmentation based on a CHM
 chmSegmentationFT(x = rasterobj,  "nameofSAGAFile")
}


}
\author{
Chris Reudenbach
}
