% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_crownSegmentation.R
\name{chmSegmentationFT}
\alias{chmSegmentationFT}
\title{fast and straightforward watershed segmentation based on 'ForestTools'}
\usage{
chmSegmentationFT(treePos = NULL, chm = NULL, minTreeAlt = 2,
  format = "polygons", verbose = FALSE)
}
\arguments{
\item{treePos}{\code{SpatialPointsDataFrame}. The point locations of treetops. The function will generally produce a
number of crown segments equal to the number of treetops.}

\item{chm}{Canopy height model in \code{raster} format. Should be the same that was used to create
the input for \code{treePos}.}

\item{minTreeAlt}{numeric. The minimum height value for a \code{CHM} pixel to be considered as part of a crown segment.
All \code{chm} pixels beneath this value will be masked out. Note that this value should be lower than the minimum
height of \code{treePos}.}

\item{format}{string. Format of the function's output. Can be set to either 'raster' or 'polygons'.}

\item{verbose}{quiet (1)}
}
\description{
'ForestTools' segmentation of individual tree crowns based on a canopy height model and initial seeding points (trees). Very fast algorithm based on the imagr watershed algorithm.
Andrew Plowright: R package \href{https://CRAN.R-project.org/package=ForestTools}{'ForestTools'}
}
\examples{
\dontrun{
 crownsFT <- chmSegmentationFT(chm = kootenayCHM,
                                 treePos = tpos 
                                format = "polygons", 
                                minTreeAlt = 1.5, 
                                verbose = FALSE)
                                
}
}
