% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_crownSegmentation.R
\name{chmSegmentationRL}
\alias{chmSegmentationRL}
\title{watershed segmentation based on rLiDAR}
\usage{
chmSegmentationRL(treePos = NULL, chm = NULL, maxCrownArea = 150,
  exclusion = 0.2)
}
\arguments{
\item{treePos}{numeric. \code{matrix} or \code{data.frame} with three columns (tree xy coordinates and height).
number of crown segments equal to the number of treetops.}

\item{chm}{Canopy height model in \link[raster]{raster} or \link[raster]{SpatialGridDataFrame} file format. Should be the same that was used to create
the input for \code{treePos}.}

\item{maxCrownArea}{numeric. A single value of the maximum individual tree crown radius expected. Default 10.0 m.
height of \code{treePos}.}

\item{exclusion}{numeric. A single value from 0 to 1 that represents the percent of pixel exclusion.}
}
\description{
'rLiDAR' segmentation of individual tree crowns based on a canopy height model and initial seeding points (trees). Generic segmentation algorithm
Carlos A. Silva et all.: R package \href{https://CRAN.R-project.org/package=rLiDAR}{rLiDAR}\cr
}
\examples{
\dontrun{
 crownsRL <- chmSegmentationRL(chm, treePos, maxCrownArea, exclusion)
}
}
