% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classificationStats.R
\name{classificationStats}
\alias{classificationStats}
\title{calculate prediction performance statistics for classification models}
\usage{
classificationStats(prd, obs, prob = NULL, plot = FALSE)
}
\arguments{
\item{prd}{factor vector of predicted values with two levels}

\item{obs}{factor vector of observed values with two levels}

\item{prob}{optional. Predicted probabilities for the first class}

\item{plot}{logical, whether to produce a visualisation of the results.
Defaults to FALSE}
}
\value{
If \code{plot = FALSE} (the default), a data frame. 
If \code{plot = TRUE}, a list with components \code{stats} - data frame
and \code{plot} - a trellis plot object.
}
\description{
this function calculates prediction performance statistics
between vectors of predicted and observed values. Users may also create 
a dotplot visualising the results.
}
\examples{
#create two random vectors with classes "yes" and "no" to simulate a model
#with random performance. Expected POD and PFD  
pred_vals <- factor(sample(c("Yes","No"), 50, replace = TRUE),levels=c("Yes","No"))
obs_vals <- factor(sample(c("Yes","No"), 50, replace = TRUE),levels=c("Yes","No"))

result <- classificationStats(pred_vals, obs_vals, plot=TRUE)
result$plot
result$stats

}
\seealso{
\code{\link{regressionStats}}
}
\author{
Hanna Meyer and Tim Appelhans
}
