% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyTreeCrown.R
\name{classifyTreeCrown}
\alias{classifyTreeCrown}
\title{calcualte and post-classifies the morphological structure of raw tree crowns}
\usage{
classifyTreeCrown(runDir,currentP,allP)
}
\arguments{
\item{crownFn}{filname of OGR comliant vector file}

\item{funNames}{morphological parameters to be calculated}

\item{minTreeAlt}{minimum height in meter that will be regarded as tree}

\item{crownMinArea}{minimum area of crowns that is accepted}

\item{crownMaxArea}{maximum area of crowns that is accepted}

\item{solidity}{minimum solidity of crowns that is accepted}

\item{WLRatio}{minimum WLRatio of crowns that is accepted}
}
\value{
classifyTreeCrown basically returns SPDF  with the crown polygons and all calculated parameters
}
\description{
calcualte and post-classifies the morphological structure of raw tree crowns
}
\examples{
#### Example to use classifyTreeCrown for a common analysis of the
    estimated spreading distances of an specifified area

#
  trees_crowns <- classifyTreeCrown(crownFn = paste0(pd_gi_run,"crownsHeight.shp"),segType = 1, 
                                      funNames = c("eccentricityboundingbox","solidity"),
                                      minTreeAlt = 5, 
                                      crownMinArea = 3, 
                                      crownMaxArea = 300, 
                                      solidity = 1, 
                                      WLRatio = 0.5)

}
\author{
Chris Reudenbach
}
