% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_bluedjinn.R
\name{convert2SAGA}
\alias{convert2SAGA}
\title{split multiband tif to single tif files}
\usage{
convert2SAGA(fn = NULL, bandName = NULL, startBand = 1, endBand = 3,
  refFn = NULL)
}
\arguments{
\item{fn}{character. filename}

\item{startBand}{numerical. first band to export}

\item{refFn}{character. reference image for resampling}

\item{bandname}{character. list of bandnames c("red","green","blue")}

\item{startBand}{numerial. last band to export}
}
\description{
split multiband tif to single tif files
}
\keyword{internal}
