% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costMergePeaks.R
\name{costMergePeaks}
\alias{costMergePeaks}
\title{Trys to merge indenpendent peak positions to provide a reliable join using a cost analysis}
\usage{
costMergePeaks(dem.peaklist,ext.peaklist)
}
\arguments{
\item{dem.peaklist}{DEM derived peaklist containing coords and altitude}

\item{ext.peaklist}{external peaklist, containing coords name and altitude}

\item{dem}{DEM data as a raster object}
}
\value{
costMergePeaks returns one merged peaklist with the names of the external peaks and the coords and altitude from the DEM derived list
}
\description{
http://moc.environmentalinformatics-marburg.de/doku.php?id=courses:msc:advanced-gis:description
}
\examples{
  
#### Example merges the two peak lists
note the ext.peaklist must contain at least valid coordinates and names

      new.peaklist<-costmergePeaks(dem.peaklist,ext.peaklist)
}
\author{
Chris Reudenbach
}
\references{
\url{http://moc.environmentalinformatics-marburg.de/doku.php?id=courses:msc:advanced-gis:description}
}

