% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_fusionTools.R
\name{createGroundSurface}
\alias{createGroundSurface}
\title{Fusion wrapper for creating a DEM raster* object from LiDAR  point clouds}
\usage{
createGroundSurface(lasFiles, res)
}
\arguments{
\item{lasFiles}{default is \code{NULL} list of the las file(s)}

\item{res}{resolution for raster operations}
}
\description{
Create a DEM raster* object from a LiDAR generated point cloud based on regular las format files
}
\examples{
\dontrun{
# create a DSM based on a uav point cloud
createGroundSurface(lasFiles = lasFiles, res=10)
}

}
\author{
Chris Reudenbach,Jannis Gottwald
}
