% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_bluedjinn.R
\name{cutTif}
\alias{cutTif}
\title{clips a tif files according to a given extent.}
\usage{
cutTif(rasterFiles = NULL, ext = NULL, outPath = "cut",
  prefix = "cut")
}
\arguments{
\item{rasterFiles}{character. vector containing a list of rasterfiles to be clipped}

\item{ext}{extent}

\item{outPath}{character. subfolder of current runtime folder. clipped files will be stored there}

\item{prefix}{character. prefic string that is added to the filenames}
}
\description{
clips a tif files according to a given extent
}
\keyword{internal}
