% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_fusionTools.R
\name{densityMetrics}
\alias{densityMetrics}
\title{Create height specific proportions of points and canopy height model as dtm* object from a LiDAR generated point cloud}
\usage{
densityMetrics(lasFiles, heightClassList, res)
}
\arguments{
\item{lasFiles}{default is \code{NULL} list of the las file(s)}

\item{heightClassList}{default is \code{list(c(0,2,5,10,30))}}

\item{res}{resolution for raster operations}
}
\description{
Create height specific proportions of points and canopy height model as dtm* object from a LiDAR generated point cloud
}
\examples{
\dontrun{
# create proportions, all returns, canopy height model in dtm
densityMetrics(lasFiles = lasFiles, heightClassList =heights, res = 10 )
}


}
\author{
Chris Reudenbach,Jannis Gottwald
}
