% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_classificationStats.R
\name{envinmr.theme}
\alias{envinmr.theme}
\title{Environmental Informatics Marburg lattice plotting theme}
\usage{
envinmr.theme(win.fontfamily = NULL, with.bg = FALSE, box = "black", ...)
}
\arguments{
\item{win.fontfamily}{character font family.}

\item{with.bg}{logical. If \code{FALSE}, the background is transparent.}

\item{box}{colour of the box around the main plot and, if present, the color key}

\item{...}{Further arguments passed on to \code{\link{simpleTheme}}.}
}
\description{
This theme is a modified version of \code{\link{theEconomist.theme}}.
Especially the regions colour setting has been modified to a
conceptually pleasant rainbow colour palette (optimised for temperature).
}
\examples{
library(lattice)
levelplot(matrix(rnorm(10000), 100, 100),
          par.settings = envinmr.theme(), at = seq(-5, 5, 0.1))

}
\seealso{
\code{\link{theEconomist.theme}}
}
\author{
Tim Appelhans
}
