% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_classify.R
\name{extractTrainData}
\alias{extractTrainData}
\title{extracts training data from a raster stack}
\usage{
extractTrainData(rasterStack = NULL, trainPlots = NULL,
  trainDataFn = filepath(temp(), "trainingDF.RData"))
}
\arguments{
\item{rasterStack}{default is \code{NULL} rasterstack wcontaining all image data}

\item{trainPlots}{default is \code{NULL}  sp object providing training geometries}

\item{trainDataFn}{default is \code{filepath(temp(),"trainingDF.RData")} Name of the extracted training data file}

\item{ids}{default is \code{c(1,2)} classification ids}

\item{idLabel}{default is \code{c("yes","no")} names of ids}
}
\value{
extractTrainData returns a dataframe with all training data
}
\description{
extracts training data from a raster stack
}
\examples{
\dontrun{

trainingDF <- extractTrainData(rasterStack  = trainStack,
                               training     = training,
                               ids=c(1,2),
                               idLabel= c("green","nogreen")) 
}

}
\author{
Chris Reudenbach
}
