% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_crownSegmentation.R
\name{fa_crown_segmentation}
\alias{fa_crown_segmentation}
\title{Tree segmentation based on a CHM}
\usage{
fa_crown_segmentation(seeds = "seeds.sgrd", is3_leafsize = 8,
  is3_normalize = 1, is3_neighbour = 0, is3_method = 0, is3_sig1 = 0.05,
  is3_sig2 = 0.05, is3_threshold = 5e-05, is3_seed_params = c("chm"),
  majority_radius = 5, seeding = TRUE, split = TRUE)
}
\arguments{
\item{seeds}{spatial raster object}

\item{is3_leafsize}{default is 8,}

\item{is3_normalize}{default is 1,}

\item{is3_neighbour}{default is 0,}

\item{is3_method}{default is 0,}

\item{is3_sig1}{default is  0.05,}

\item{is3_sig2}{default is  0.05,}

\item{is3_threshold}{default is  0.0001,}

\item{is3_seed_params}{default is c("GLI","HI","GRV") rgb image derived indices}

\item{seeding}{default  is TRUE switch if seeding is called}

\item{split}{default  is TRUE switch if splitting of the polygons is called}
}
\value{
basically returns a  vector data sets with the tree crown geometries and a bunch of corresponding indices
}
\description{
Tree segmentation based on a CHM
}
\examples{
\dontrun{
# Tree segmentation based on a CHM
 fa_crown_segmentation(x = rasterobj,  "nameofSAGAFile")
}

}
\author{
Chris Reudenbach
}
