% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_crownSegmentation.R
\name{fa_crownSegementationDP}
\alias{fa_crownSegementationDP}
\title{watershed segmentation based on itcSegment}
\usage{
fa_crownSegementationDP(chm = NULL, EPSG_code = 3064, mov_window = 7,
  TRESHSeed = 0.45, TRESHCrown = 0.55, maxTreeAlt = 2,
  maxCrownArea = 100)
}
\arguments{
\item{chm}{Canopy height model in \link[raster]{raster} or \link[raster]{SpatialGridDataFrame} file format. Should be the same that was used to create
the input for \code{treePos}.}

\item{mov_window}{Size (in pixels) of the moving window to detect local maxima.}

\item{maxCrownArea}{numeric. A single value of the maximum individual tree crown radius expected. Default 10.0 m.
height of \code{treePos}.}

\item{treePos}{numeric. \code{matrix} or \code{data.frame} with three columns (tree xy coordinates and height).
number of crown segments equal to the number of treetops.}

\item{EPSG}{The EPSG code of the reference system of the CHM raster image.}

\item{th_local_max}{Height threshold (m) below a pixel cannot be a local maximum. Local maxima values are used to define tree tops.}
}
\description{
Segmentation of individual tree crowns as polygons based on a LiDAR derived canopy height model. 
Michele Dalponte: R package \href{https://CRAN.R-project.org/package=itcSegment}{itcSegment}\cr
M. Dalponte, F. Reyes, K. Kandare, and D. Gianelle, "Delineation of Individual Tree Crowns from ALS and Hyperspectral data: a comparison among four methods," European Journal of Remote Sensing, Vol. 48, pp. 365-382, 2015.
}
\examples{
\dontrun{
 crownsRL <- fa_crownSegementationLR(chm, treePos, maxCrownArea, exclusion)
}
}
