% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_crownSegmentation.R
\name{fa_crownSegmentation}
\alias{fa_crownSegmentation}
\title{Tree segmentation based on a CHM}
\usage{
fa_crownSegmentation(treePos = NULL, chm = NULL, minTreeAlt = 2,
  mintreeAltParam = "chmQ20", is3_leafsize = 8, is3_normalize = 1,
  is3_neighbour = 0, is3_method = 0, is3_thVarFeature = 0.05,
  is3_thVarSpatial = 0.05, is3_thSimilarity = 5e-05,
  is3_seed_params = c("chm"), majority_radius = 2, giLinks = NULL)
}
\arguments{
\item{treePos}{spatial raster object}

\item{chm}{Canopy height model in \link[raster]{raster} format. Should be the same that was used to create
the input for \code{treePos}.}

\item{minTreeAlt}{numeric. The minimum height value for a \code{chm} pixel to be considered as part of a crown segment.
All \code{chm} pixels beneath this value will be masked out. Note that this value should be lower than the minimum
height of \code{treePos}.}

\item{mintreeAltParam}{default is "chmQ20"}

\item{is3_leafsize}{integer, bin size of grey value sampling range from 1 to 256 default is 8,}

\item{is3_normalize}{integer,  logical switch if data will be normalized or not default is 1,}

\item{is3_neighbour}{integer,  von Neumanns' neighborhood (0) or Moore's (1) default is 0,}

\item{is3_method}{integer, growing algorithm for feature space and position (0) or feature space only (1)}

\item{is3_thVarFeature}{numerical, spatial variance default is  0.05,}

\item{is3_thVarSpatial}{numerical, spatial variance default is  0.05}

\item{is3_thSimilarity}{mumerical similarity threshold default is  0.00005,}

\item{is3_seed_params}{vector of characters corresponding with the used attributes default is c("chm") altitude values from surface model}

\item{giLinks}{list of GI tools cli pathes  default is NULL}
}
\description{
Tree segmentation based on a CHM, basically returns a  vector data sets with the tree crown geometries and a bunch of corresponding indices. After the segementation itself the results are hole filled and optionally filtered by a majority filter in the 3*3 surrounding.
}
\examples{
\dontrun{
# Tree segmentation based on a CHM
 fa_crownSegmentation(x = rasterobj,  "nameofSAGAFile")
}


}
\author{
Chris Reudenbach
}
