% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_crownSegmentation.R
\name{fa_crownSegmentationFT}
\alias{fa_crownSegmentationFT}
\title{very fast watershed segemenation algorithm provided by 'ForestTools'}
\arguments{
\item{treePos}{\link[sp]{SpatialPointsDataFrame}. The point locations of treetops. The function will generally produce a
number of crown segments equal to the number of treetops.}

\item{chm}{Canopy height model in \link[raster]{raster} format. Should be the same that was used to create
the input for \code{treePos}.}

\item{minTreeAlt}{numeric. The minimum height value for a \code{CHM} pixel to be considered as part of a crown segment.
All \code{chm} pixels beneath this value will be masked out. Note that this value should be lower than the minimum
height of \code{treePos}.}

\item{format}{string. Format of the function's output. Can be set to either 'raster' or 'polygons'.}
}
\description{
Segmentation of individual tree crowns based on a canopy height model and initial seeding points (trees). Very fast algorithm based on the imagr watershed algorithm.
Andrew Plowright: R package \href{https://CRAN.R-project.org/package=ForestTools}{'ForestTools'}
}
\examples{
\dontrun{
 crownsFT <- crownSegementationFT(chm = kootenayCHM,
                                 treePos = tpos 
                                format = "polygons", 
                                minTreeAlt = 1.5, 
                                verbose = FALSE)
}
}
