% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_treePosition.R
\name{fa_findTreePosition}
\alias{fa_findTreePosition}
\title{Tree segmentation based on a CHM}
\usage{
fa_findTreePosition(chm = NULL, minTreeAlt = 10,
  minTreeAltParam = "chmQ20", minCrownArea = 3, maxCrownArea = 150,
  is0_output = 1, is0_join = 1, is0_thresh = 0.1, split = TRUE,
  giLinks = NULL)
}
\arguments{
\item{chm}{spatial raster object}

\item{minTreeAlt}{default is 5}

\item{minCrownArea}{default is 3 minimum area of crown}

\item{maxCrownArea}{default is 225 maximum area of crown}

\item{is0_output}{default is 0,     # 0=s treePos value 1=segment id}

\item{is0_join}{default is 2,     # 0=no join, 1=treePos2saddle diff, 2=treePos2treePos diff}

\item{is0_thresh}{default is 0.05,  # threshold for join difference in m}

\item{split}{default  is TRUE switch if splitting of the polygons is called}

\item{mintreeAltParam}{default is "chmQ20"}
}
\value{
basically returns a  vector data sets with the tree crown geometries and a bunch of corresponding indices
}
\description{
Tree segmentation based on a CHM
}
\examples{
\dontrun{
# Tree segmentation based on a CHM
 fa_findTreePosition(chm = rasterobj,  "nameofSAGAFile")
}

}
\author{
Chris Reudenbach
}
