% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_pot_insolation.R
\name{fa_pot_insol}
\alias{fa_pot_insol}
\title{calculate statitiscs of polygon based raster extraction}
\usage{
fa_pot_insol(x = NULL, svf_radius = 10000, svf_method = 0,
  svf_dlevel = 3, svf_ndirs = 8, pi_solarconst = 1367, pi_units = 0,
  pi_shadow = 1, pi_day = "21/06/17", pi_day_stop = "21/06/17",
  pi_day_step = 1, pi_hour_range_min = 0, pi_hour_range_max = 24,
  pi_hour_step = 0.5, pi_lumped = 70)
}
\arguments{
\item{x}{spatial raster object}

\item{svf_radius}{search radius in map units default is 10000.0, set to zero for ignoring,}

\item{svf_dlevel}{multi scale factor default is 3.0 minimum is 1.25}

\item{pi_solarconst}{(W/m**2) default is 1367.0,}

\item{pi_units}{units for output radiation values. default is 0 (kWh/m**2) available are: 1 (kJ/m**2) 2 J/cm**2,}

\item{pi_shadow}{default is 1 (fat), choose 0 (slim) to trace grid node's shadow, 1 (fat) to trace the whole cell's shadow, or 2 (none) to ignore shadowing effects. The first is slightly faster but might show some artifacts}

\item{pi_day}{= "21/06/17",}

\item{pi_day_stop}{= "21/06/17",}

\item{pi_day_step}{= 1}

\item{pi_hour_range_min}{= 0.0,}

\item{pi_hour_range_max}{= 24.0,}

\item{pi_hour_step}{= 1.0,}

\item{pi_lumped}{= 70.0}

\item{sfv_method}{default is 0 (multiscale) NOTE much slower than 1 (sectors),}

\item{sfv_ndirs}{number of sectors default is 8.0 minimum is 3.0}
}
\value{
returns all calculations as SAGA Grid for statistical analysis
}
\description{
calculate statitiscs of polygon based raster extraction
}
\examples{
\dontrun{
# potential insolation based on a CHM
pot_insolation <- fa_pot_insol("chm")
}

}
\author{
Chris Reudenbach
}
