% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_treeSegementation.R
\name{fa_tree_segementation}
\alias{fa_tree_segementation}
\title{Tree segementation based on a CHM}
\usage{
fa_tree_segementation(x = NULL, minTreeAlt = 3, crownMinArea = 7,
  crownMaxArea = 200, WLRatio = 0.53, thLongit = 0.5, solidity = 1,
  is0_output = 1, is0_join = 1, is0_thresh = 0.09, is3_leafsize = 8,
  is3_normalize = 1, is3_neighbour = 1, is3_method = 0, is3_sig1 = 0.1,
  is3_sig2 = 3.01, is3_threshold = 0.001, is3_param2 = "GLI",
  is3_param1 = "HI", majority_radius = 5)
}
\arguments{
\item{x}{spatial raster object}

\item{minTreeAlt}{= 3,    # -thresholdfor minimum tree altitude in meter}

\item{crownMinArea}{= 7,    #(approx 1.25 m diameter)}

\item{crownMaxArea}{= 200,  #(approx 17.8 m diameter)}

\item{WLRatio}{= 0.53, # crown width length ratio}

\item{thLongit}{= 0.5,  # crown longitudiness}

\item{solidity}{= 1.0, # solidity}

\item{is0_output}{= 0,     # 0= seed value 1=segment id}

\item{is0_join}{= 2,     # 0=no join, 1=seed2saddle diff, 2=seed2seed diff}

\item{is0_thresh}{= 0.05,  # threshold for join difference in m}

\item{is3_leafsize}{= 8,}

\item{is3_normalize}{= 1,}

\item{is3_neighbour}{= 1,}

\item{is3_method}{= 0,}

\item{is3_sig1}{=  0.1,}

\item{is3_sig2}{= 3.01,}

\item{is3_threshold}{= 0.001,}

\item{is3_param2}{default is HI  GLI next rgb image derived index}

\item{is3_param1}{default is HI first rgb image derived index}

\item{majority_radius}{= 5.000}
}
\value{
basically returns a  vector data sets with the tree crown geometries and a bunch of corresponding indices
}
\description{
Tree segementation based on a CHM
}
\examples{
\dontrun{
# Tree segementation based on a CHM
 fa_tree_segementation(x = rasterobj,  "nameofSAGAFile")
}

}
\author{
Chris Reudenbach
}
