% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_treeSeeding.R
\name{fa_treeSeeding}
\alias{fa_treeSeeding}
\title{Tree segmentation based on a CHM}
\usage{
fa_treeSeeding(x = NULL, minTreeAlt = 10, crownMinArea = 3,
  crownMaxArea = 150, is0_output = 1, is0_join = 1, is0_thresh = 0.1,
  split = TRUE)
}
\arguments{
\item{x}{spatial raster object}

\item{crownMinArea}{default is 3 minimum area of crown}

\item{is0_output}{default is 0,     # 0=s seed value 1=segment id}

\item{is0_join}{default is 2,     # 0=no join, 1=seed2saddle diff, 2=seed2seed diff}

\item{is0_thresh}{default is 0.05,  # threshold for join difference in m}

\item{split}{default  is TRUE switch if splitting of the polygons is called}

\item{minTreealt}{default is 5}

\item{crownMinArea}{default is 225 maximum area of crown}
}
\value{
basically returns a  vector data sets with the tree crown geometries and a bunch of corresponding indices
}
\description{
Tree segmentation based on a CHM
}
\examples{
\dontrun{
# Tree segmentation based on a CHM
 fa_treeSeeding(x = rasterobj,  "nameofSAGAFile")
}

}
\author{
Chris Reudenbach
}
