% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_bluedjinn.R
\name{fileProcStatus}
\alias{fileProcStatus}
\title{returns the status of saga modules}
\usage{
fileProcStatus(module = NULL, file = NULL, listname = NULL)
}
\arguments{
\item{module}{character name of module to be checked}

\item{file}{character. filename to be checked}

\item{listname}{character. name of log list}
}
\description{
returns the status of saga modules
}
\keyword{internal}
