% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foa_treeSegementation.R
\name{foa_tree_segementation}
\alias{foa_tree_segementation}
\title{Tree segementation based on a CHM}
\usage{
foa_tree_segementation(chmFn = NULL, minTreeAlt = 3, thtreeNodes = 6,
  crownMinArea = 7, crownMaxArea = 200, WLRatio = 0.53, thLongit = 0.5,
  solidity = 1, is0_output = 0, is0_join = 2, is0_thresh = 0.05,
  is3_leafsize = 8, is3_normalize = 1, is3_neighbour = 1,
  is3_method = 0, is3_sig1 = 0.1, is3_sig2 = 3.01,
  is3_threshold = 0.001, majority_radius = 5)
}
\arguments{
\item{chmFn}{filname auf SAGA grid}

\item{minTreeAlt}{= 3,    # -thresholdfor minimum tree altitude in meter}

\item{thtreeNodes}{= 6,    # minimum number of ldd connections}

\item{crownMinArea}{= 7,    #(approx 1.25 m diameter)}

\item{crownMaxArea}{= 200,  #(approx 17.8 m diameter)}

\item{WLRatio}{= 0.53, # crown width length ratio}

\item{thLongit}{= 0.5,  # crown longitudiness}

\item{solidity}{= 1.0, # solidity}

\item{is0_output}{= 0,     # 0= seed value 1=segment id}

\item{is0_join}{= 2,     # 0=no join, 1=seed2saddle diff, 2=seed2seed diff}

\item{is0_thresh}{= 0.05,  # threshold for join difference in m}

\item{is3_leafsize}{= 8,}

\item{is3_normalize}{= 1,}

\item{is3_neighbour}{= 1,}

\item{is3_method}{= 0,}

\item{is3_sig1}{=  0.1,}

\item{is3_sig2}{= 3.01,}

\item{is3_threshold}{= 0.001,}

\item{majority_radius}{= 5.000}
}
\value{
basically returns a  vector data sets with the tree crown geometries and a bunch of corresponding indices
}
\description{
Tree segementation based on a CHM
}
\examples{
\dontrun{
# Tree segementation based on a CHM
 foa_tree_segementation(chmFn =  "nameofSAGA.grd")
}

}
\author{
Chris Reudenbach
}
