% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_fusionTools.R
\name{fu_sliceRas}
\alias{fu_sliceRas}
\title{Fusion wrappper to create a raster* object from a LiDAR point cloud}
\usage{
fu_sliceRas(lasFiles = NULL, zrange = NULL, zrnames = NULL,
  paramList = c("1 M M 1 32 0 0 "), res = 10,
  proj4 = "+proj=utm +zone=32 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs")
}
\arguments{
\item{lasFiles}{default is \code{NULL} list of the las file(s)}

\item{zrange}{default is \code{list(c(0,3), c(3,15))} list vector pairs describing the vertical slice}

\item{zrnames}{zrnames}

\item{paramList}{default is \code{c("1 M M 1 32 0 0 ")}list of parameters for info see clipdata.exe manual}

\item{res}{resolution for raster operations}

\item{proj4}{any valid proj4 string that is assumingly the correct one default is \code{"+proj=utm +zone=32 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs"}}
}
\description{
Create a raster* object from a LiDAR generated point cloud based on regular las format files
}
\examples{
\dontrun{
# create a DSM based on a uav point cloud
r<-lasSlice(lasFiles = lasFiles)
}zrange[j],"_",zrange[j+1]

}
\author{
Chris Reudenbach,Jannis Gottwald
}
