% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_karim.R
\name{G2OGR}
\alias{G2OGR}
\title{converts GRASS vector to shape file}
\usage{
G2OGR(runDir = NULL, layer = NULL)
}
\arguments{
\item{runDir}{path of working directory}

\item{layer}{name GRASS raster}
}
\description{
converts GRASS vector to shape file
}
