% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_karim.R
\name{G2Tiff}
\alias{G2Tiff}
\title{converts GRASS raster to geotiff}
\usage{
G2Tiff(runDir = NULL, layer = NULL, returnRaster = FALSE)
}
\arguments{
\item{runDir}{path of working directory}

\item{layer}{name GRASS raster}

\item{returnRaster}{return GRASS raster as an R raster object default = FALSE}
}
\description{
converts GRASS raster to geotiff
}
