% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gCost.R
\name{gcost}
\alias{gcost}
\title{calculates the least cost distance using an existing cost raster.}
\usage{
gcost(runDir,currentP,allP)
}
\arguments{
\item{runDir}{the choosen working directory}

\item{currentP}{actual startpoint}

\item{allP}{list of all Points}
}
\value{
gcost basically returns (a) dataframe with the position of startpoint,destination point and distance. Additionally it writes a shapefile containing all pathes to the destinations for the source point
}
\description{
least cost path algorithm are useful for al lot of applications. \link{gdistance} is providing a generic R-ish approach that is cool for smaller data set. If you want to go bigger than 2000*2000 pixel than you should engage something else. This implementation uses the incredible power of GRASS 7. It calculates the least cost pathes for a set of given points for each point in the list to all other points.
}
\examples{
#### Example to use gcost for a common analysis of the
    estimated spreading distances of an specifified area

#
gcost(rootDir = "/home/creu/proj/beetle/cost",c("91.52", "30.62"),c("91.28" ,"30.1"))

}
\author{
Chris Reudenbach
}

