% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_textures.R
\name{gdal_dem}
\alias{gdal_dem}
\title{calculate gdal derived DEM params}
\usage{
gdal_dem(dem, item = NULL)
}
\arguments{
\item{dem}{of GeoTiff containing one channel DEM}

\item{item}{index to be calculated default are c("slope", "aspect","TRI","TPI","Roughness")}
}
\description{
calculate gdal derived DEM params
}
\note{
please provide a GeoTiff file
}
\examples{
\dontrun{
url<-"http://www.ldbv.bayern.de/file/zip/5619/DOP\%2040_CIR.zip"
res <- curl::curl_download(url, "testdata.zip")
unzip(res,junkpaths = TRUE,overwrite = TRUE)
gm<-gdal_dem(dem=paste0(getwd(),"4490600_5321400.tif"))
raster::plot(gm[[1]])
}
}
