% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_classify.R
\name{get_counts}
\alias{get_counts}
\title{counts pixel values according to their classes}
\usage{
get_counts(ids = c(1, 2), position = NULL, imageFiles = NULL,
  buffersize = 1.5, out_prefix = "classified_index_", ext = ".tif",
  path = path_output, dropChars = 0)
}
\arguments{
\item{ids}{ids}

\item{position}{position}

\item{imageFiles}{image files}

\item{buffersize}{size in meters around position}

\item{out_prefix}{out prefix string}

\item{ext}{extension}

\item{path}{output path}

\item{dropChars}{chars to drop}
}
\description{
counts pixel values according to their classes
}
\examples{
\dontrun{
df1<-get_counts(position = position,
              imageFiles = imageFiles,
              dropChars = 8,
              pre=pre,
              ext=".tif")
}
}
