% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_bluedjinn.R
\name{get_gi}
\alias{get_gi}
\title{creates all link2GI links}
\usage{
get_gi(links = NULL, simple = TRUE, linkItems = c("saga", "grass7", "otb",
  "gdal"), sagaArgs = "default", grassArgs = "default",
  otbArgs = "default",
  gdalArgs = "quiet = TRUE,\\n                                     returnPaths = TRUE")
}
\arguments{
\item{links}{character links}

\item{simple}{boolean true  make all}

\item{linkItems}{character list of c("saga","grass7","otb","gdal")}

\item{sagaArgs}{sagaArgs}

\item{grassArgs}{grassArgs}

\item{otbArgs}{otbArgs}

\item{gdalArgs}{gdalArgs}
}
\description{
converts SAGA raster to R raster object
}
