% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_karim.R
\name{get_max_posFromPoly}
\alias{get_max_posFromPoly}
\title{extract for all polygons the position of the maximum value of the applied raster(s)}
\usage{
get_max_posFromPoly(dem, lN, poly_split = TRUE)
}
\description{
extract for all polygons the position of the maximum value
}
