% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_bluedjinn.R, R/fun_getCrayon.R
\name{getCrayon}
\alias{getCrayon}
\title{colorize the cat outputs}
\usage{
getCrayon()

getCrayon()
}
\description{
colorize the cat outputs

colorize the cat outputs
}
\keyword{internal}
