% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_karim.R
\name{getmaxposFromLine}
\alias{getmaxposFromLine}
\title{applies a line to a raster and returns the position of the maximum value}
\usage{
getmaxposFromLine(dem, line)
}
\arguments{
\item{dem}{raster object}

\item{line}{sp object}
}
\description{
applies a line to a raster and returns the position of the maximum value
}
