% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_bluedjinn.R
\name{grass2shape}
\alias{grass2shape}
\title{converts GRASS vector to shape file}
\usage{
grass2shape(runDir = NULL, layer = NULL)
}
\arguments{
\item{runDir}{path of working directory}

\item{layer}{name GRASS raster}
}
\description{
converts GRASS vector to shape file
}
