% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_bluedjinn.R
\name{grass2tif}
\alias{grass2tif}
\title{converts GRASS raster to Geotiff}
\usage{
grass2tif(runDir = NULL, layerName = NULL, returnRaster = FALSE)
}
\arguments{
\item{runDir}{path of working directory}

\item{layerName}{name GRASS raster}

\item{returnRaster}{return GRASS raster as an R raster object, default = FALSE}
}
\value{
if returnRaster the converted GRASS raster is returned as an raster* object
}
\description{
converts GRASS raster to Geotiff
}
\keyword{internal}
