% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_fusionTools.R
\name{gridMetrics}
\alias{gridMetrics}
\alias{GridMetrics}
\title{calculate elevation and intensity based metrics an write out selected metrics to raster (.asc)}
\usage{
GridMetrics(lasFiles, res, heightClassList, metrics, heightbreak, buffer,
  cellbuffer, strata, instrata,
  proj4 = "+proj=utm +zone=32 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs")
}
\arguments{
\item{lasFiles}{default is \code{NULL} list of the las file(s)}

\item{res}{default is \code{NULL} cellsize}

\item{heightClassList}{height classes list default is \code{list(c(0,2,5,10,30))}}

\item{metrics}{default is  \code{list(c(0,2,5,10,30))} list of metric indexes indexes given
in Fusion Manual p.78/79}

\item{heightbreak}{lower heightbreak}

\item{buffer}{optional parameter \code{NULL} if extent should be extendet by a buffer}

\item{cellbuffer}{optional parameter \code{NULL} if cellbased metrics should
take neighbouring cells into account}

\item{strata}{optional paramter \code{list(c(0,2,5,10,30))}, elevation based metrics
will be calculated for strata}

\item{instrata}{optional paramter \code{list(c(0,2,5,10,30))}, intensity based metrics
will be calculated for strata}

\item{proj4}{any valid proj4 string that is assumingly the correct one default is \code{"+proj=utm +zone=32 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs"}}
}
\description{
calculate elevation and intensity based metrics an write out selected metrics to raster (.asc)
}
\examples{
\dontrun{

GridMetrics(lasFiles = lasFiles, 
                res = res, 
    heightClassList = h, 
    metrics = 31, 
    heightbreak=0.2 )
}


}
\author{
Chris Reudenbach,Jannis Gottwald
}
