% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_karim.R
\name{h_grass2shape}
\alias{h_grass2shape}
\title{converts GRASS vector to shape file}
\usage{
h_grass2shape(runDir = NULL, layer = NULL)
}
\arguments{
\item{runDir}{path of working directory}

\item{layer}{name GRASS raster}
}
\description{
converts GRASS vector to shape file
}
