% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_karim.R
\name{h_line_extract_maxpos}
\alias{h_line_extract_maxpos}
\title{applies a line to a raster and returns the position of the maximum value}
\usage{
h_line_extract_maxpos(dem, line)
}
\arguments{
\item{dem}{raster object}

\item{line}{sp object}
}
\description{
applies a line to a raster and returns the position of the maximum value
}
