% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_karim.R
\name{h_r2saga}
\alias{h_r2saga}
\title{converts SAGA raster to R raster object}
\usage{
h_r2saga(x, fn)
}
\arguments{
\item{x}{raster object}

\item{fn}{filname without extension}
}
\description{
converts SAGA raster to R raster object
}
