% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_karim.R
\name{h_read_gpx}
\alias{h_read_gpx}
\title{Read GPX file}
\usage{
h_read_gpx(file, layers = c("waypoints", "tracks", "routes", "track_points",
  "route_points"))
}
\arguments{
\item{file}{a GPX filename (including directory)}

\item{layers}{vector of GPX layers. Possible options are \code{"waypoints"}, \code{"tracks"}, \code{"routes"}, \code{"track_points"}, \code{"route_points"}. By dedault, all those layers are read.}
}
\value{
if the layer has any features a sp object is returned.
}
\description{
Read a GPX file. By default, it reads all possible GPX layers, and only returns shapes for layers that have any features.
}
\note{
cloned from tmap
}
