% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_karim.R
\name{h_saga2r}
\alias{h_saga2r}
\title{converts SAGA raster to R raster object}
\usage{
h_saga2r(fn, ext)
}
\arguments{
\item{fn}{filname without extension}

\item{ext}{extent of the raster in R notation}
}
\description{
converts SAGA raster to R raster object
}
