% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_karim.R
\name{h_shape2grass}
\alias{h_shape2grass}
\title{converts OGR to GRASS vector}
\usage{
h_shape2grass(runDir = NULL, layer = NULL)
}
\arguments{
\item{runDir}{path of working directory}

\item{layer}{name GRASS raster}
}
\description{
converts OGR to GRASS vector
}
