% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initRGIS.R
\name{initRGIS}
\alias{initRGIS}
\title{Initializes RSAGA, rgrass7 and gdalUtil}
\usage{
initRGIS(fname.control,)
}
\arguments{
\item{root.dir}{project directory}

\item{working.dir}{runtime directory}

\item{fndem}{name of a (DEM) GDAL format raster file}

\item{rasterParam}{a rasterobject link}
}
\value{
initRGIS initializes the usage of GRASS7 RSAGA 2.12+ and gdalUtil\cr
it returns the RSAGA env, the GRASS env and the gdalUtils status
}
\description{
Function that initializes environment, pathes, SAGA, GRASS7 and GDAL support via gdalUtils.
            *NOTE* you probably have to customize some settings according to your personal installation folders.
}
\details{
The concept is very straightforward but for an all days usage pretty helpful. You need to provide a GDAL conform raster file, a \link{raster} object or you may download SRTM data with \link{getGeoData}. This settings will be used to initialize a temporary but static \href{https://cran.r-project.org/web/packages/rgrass7}{rgrass7} environment. Additionally \href{https://cran.r-project.org/web/packages/RSAGA}{RSAGA} and \href{https://cran.r-project.org/web/packages/gdalUtils}{gdalUtils} are initialized and checked. During the rsession you will have full access to GRASS7 and SAGA via the wrapper packages. .
}
\examples{
#### Examples how to use RSAGA and GRASS bindings from R

# full call using getGeodata
envGIS<- initRGIS <- function(root.dir=tempdir(),working.dir='cost',fndem=getGeoData("SRTM",xtent=c(49,49,11,11)))

# initialisation with an existing raster object
envGIS<- initRGIS("~/proj/tester",'cost',fndem=NULL,rasterParam = outraster)

# init with an local GDAL file
envGIS<- initRGIS("~/proj/tester",'cost',"~/proj/runbeetle/srtm_43_01.tif")


}
\author{
Chris Reudenbach
\email{reudenbach@gis-ma.org}
}

