% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_lasTools.R
\name{lasTool}
\alias{lasTool}
\title{lasR Using a bit of lastools from R}
\usage{
lasTool(tool = "lasinfo", lasFile = NULL, thin_with_grid = "1.0",
  keep_class = "2", bulge = "1.5", step_size = "city",
  sub_size = "ultra_fine", grid_size = "1.0",
  proj4 = "+proj=utm +zone=32 +datum=WGS84 +units=m +no_defs",
  rscale = "0.01 0.01 0.01", stepoverlap = 2, cores = "2")
}
\arguments{
\item{tool}{default is \code{lasinfo}   additionally xou may choose las2las, lasmerge, lasground_new, las2dem, las2txt}

\item{lasFile}{default is \code{NULL} path  to the laz/las file(s)}

\item{thin_with_grid}{default 0.5 meter. Grid stepsize for data thinning}

\item{keep_class}{default is 2. Default ground class of las/laz conform data}

\item{bulge}{default is 1.5. 'A parameter to filter spikes it is set to a step_size/10 and then clamped into the range from 1.0 to 2.0}

\item{step_size}{default is 25 meter. LAStools key words if \code{city},\code{town},\code{metro},\code{nature},\code{wilderness} or experiment with free values}

\item{sub_size}{= "8", default is 8 meter. LAStools key words if \code{extra_coarse},\code{coarse},\code{fine},\code{extra_fine},\code{ultra_fine},\code{hyper_fine} or experiment with free values}

\item{grid_size}{resolution of the DTM raster}

\item{proj4}{default is EPSG 32632 any valid proj4 string that is assumingly the correct one}

\item{rscale}{rscale}

\item{stepoverlap}{stepoverlap}

\item{cores}{number of cores that will be used}
}
\description{
using lasLib from R. returns lasR basically returns a DTM
}
\examples{
\dontrun{
lasR(lasFile =  "~/path/to/lasdata",
       gisdbase_path = "~/temp5",
       thin_with_grid = "0.5",
       level_max = "5" ,
       grid_size = "0.5")
}
}
\author{
Chris Reudenbach
}
