% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_bluedjinn.R
\name{line_maxpos}
\alias{line_maxpos}
\title{applies a line to a raster and returns the position of the maximum value.}
\usage{
line_maxpos(dem, line)
}
\arguments{
\item{dem}{raster object}

\item{line}{sp object}
}
\value{
returns the coordinate of the maximum value of any underlying raster object along a line of interest
}
\description{
applies a line to a raster and returns the position of the maximum value
}
\keyword{internal}
