% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_bluedjinn.R
\name{linkAll}
\alias{linkAll}
\title{convenient function to establish all link2GI links}
\usage{
linkAll(links = NULL, simple = TRUE, linkItems = c("saga", "grass7",
  "otb", "gdal"), sagaArgs = "default", grassArgs = "default",
  otbArgs = "default",
  gdalArgs = c("quiet = TRUE,\\n                                     returnPaths = TRUE"))
}
\arguments{
\item{links}{character. links}

\item{simple}{logical. true  make all}

\item{linkItems}{character. list of c("saga","grass7","otb","gdal")}

\item{sagaArgs}{character. full string of sagaArgs}

\item{grassArgs}{character. grassArgs full string of grassArgs}

\item{otbArgs}{character. full string of otbArgs}

\item{gdalArgs}{character. full string of gdalArgs}
}
\description{
brute force search, find and linkl of all link2GI link functions
}
\note{
You may also use the full parameterization of the \code{link2GI} package, but you are strungly advaced to use the \code{link2GI} functions in a direct way.
}
\examples{
\dontrun{
# required packages
require(uavRst)
require(link2GI)

# search, find and create the links to all supported  GI software
giLinks<-uavRst::linkAll()

}
}
