% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_lidr_makros.R
\name{llas2llv0}
\alias{llas2llv0}
\title{checks extension mismatch of a lasfile}
\usage{
llas2llv0(las_files, level0path,
  proj4 = "+proj=utm +zone=32 +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0")
}
\arguments{
\item{level0path}{path for saving the corrected level0 data}

\item{proj4}{correct proj4 string}

\item{lasfile}{filename of an las file}
}
\description{
checks extension mismatch of a lasfile
}
