% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_classify.R
\name{make_bandnames}
\alias{make_bandnames}
\title{create name vector corresponding to the training image stack}
\usage{
make_bandnames(rgbi = NA, bandNames = NA, stat = FALSE, morpho = NA,
  edge = NA, RGBtrans = NA)
}
\arguments{
\item{rgbi}{character. codes of the RGB indices}

\item{bandNames}{character.  band names}

\item{stat}{character.  stat codes}

\item{morpho}{character.  morpho codes}

\item{edge}{character.  edge codes}

\item{RGBtrans}{character.  RGBtrans codes}
}
\description{
create name vector corresponding to the training image stack
}
\keyword{internal}
