% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_lidr_makros.R
\name{make_lidr_catalog}
\alias{make_lidr_catalog}
\title{create lidR catalog}
\usage{
make_lidr_catalog(path = NULL, cores = 1, chunksize = 0,
  chunkbuffer = 30, alignment = c(0, 0), progress = TRUE,
  stop_early = TRUE, w2w = TRUE, output_files = "", select = "",
  filter = "",
  proj4 = "+proj=utm +zone=32 +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0")
}
\arguments{
\item{path}{path for saving the corrected level0 data}

\item{cores}{number of cores for parallel procession}

\item{chunksize}{chunksiz in map units}

\item{chunkbuffer}{buffersize of chunk in mapunits}

\item{alignment}{<- value,}

\item{progress}{<- value,}

\item{stop_early}{<- value,}

\item{w2w}{<- value,}

\item{output_files}{<- value,}

\item{select}{<- value,}

\item{filter}{<- value}

\item{proj4}{correct proj4 string}

\item{cores}{<- value,}
}
\description{
create lidR catalog
}
