% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_classify.R
\name{makebNames}
\alias{makebNames}
\title{creates a parameter list from the input files}
\usage{
makebNames(rgbi = "", haratxt = "", stat = FALSE, morpho = "",
  edge = "")
}
\arguments{
\item{rgbi}{band names}

\item{haratxt}{band names}

\item{stat}{band names}

\item{morpho}{band names}

\item{edge}{band names}
}
\description{
creates a parameter list from the input files
}
