% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_merge_rgb.R
\name{merge_rgb}
\alias{merge_rgb}
\title{merge single images}
\usage{
merge_rgb(
  files = NULL,
  proj4 = NULL,
  output = NULL,
  cropoutput = NULL,
  ext = NULL
)
}
\arguments{
\item{files}{list of filenames to be merged}

\item{proj4}{valid proj4 string}

\item{output}{valid path and name for the output image}

\item{ext}{valid extent of the the area that is requested}
}
\value{
The merged image as a raster object 
# example
# tif_files is a list of valid filenames
merged_r = merge_rgb(files = tif_files,
                     out = file.path(tempdir(),"out.tif"))
}
\description{
merge single geocoded images to on big image
}
\details{
https://stackoverflow.com/questions/15876591/merging-multiple-rasters-in-r
    with regards to Flo Detsch who asked it the right way
}
\note{
if proj4 is NULL no reprojection is performed otherwise the merged raster is projected by using the proj4 string.
}
