% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_fusionTools.R
\name{missingExtents}
\alias{missingExtents}
\title{write missing extention into catalog.exe generated .csv}
\usage{
missingExtents(catalogTable)
}
\arguments{
\item{catalogTable}{default is NULL .csv generated by catalog.exe}
}
\description{
writes missing extention into catalog.exe generated .csv by substringing .las filename
}
\author{
Chris Reudenbach,Jannis Gottwald
}
