% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_karim.R
\name{OGR2G}
\alias{OGR2G}
\title{converts OGR to GRASS vector}
\usage{
OGR2G(runDir = NULL, layer = NULL)
}
\arguments{
\item{runDir}{path of working directory}

\item{layer}{name GRASS raster}
}
\description{
converts OGR to GRASS vector
}
