% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_InTex.R
\name{otbEdge}
\alias{otbEdge}
\title{Calculates edges for a given kernel size.}
\usage{
otbEdge(input = NULL, out = "edge", ram = "8192", filter = "gradient",
  filter.touzi.xradius = 1, filter.touzi.yradius = 1, channel = NULL,
  retRaster = FALSE, outDir = NULL, verbose = FALSE)
}
\arguments{
\item{input}{of GeoTiff containing 1 ore more gray value band(s)}

\item{out}{the output mono band image containing the edge features}

\item{ram}{reserved memory in MB}

\item{filter}{the choice of edge detection method (gradient/sobel/touzi)}

\item{filter.touzi.xradius}{x radius of the Touzi processing neighborhood (if filter==touzi) (default value is 1 pixel)}

\item{filter.touzi.yradius}{y radius of the Touzi processing neighborhood (if filter==touzi) (default value is 1 pixel)}

\item{channel}{sequence of bands to be processed}

\item{retRaster}{boolean if TRUE a raster stack is returned}

\item{outDir}{output Directory}

\item{verbose}{switch for system messages default is FALSE}
}
\description{
Calculates edges for a given kernel size. return list of geotiffs containing thelocal statistics for each channel
}
\note{
the otb is used for filtering. please provide a GeoTiff file
}
\examples{
\dontrun{
url<-"http://www.ldbv.bayern.de/file/zip/5619/DOP\%2040_CIR.zip"
res <- curl::curl_download(url, "testdata.zip")
unzip(res,junkpaths = TRUE,overwrite = TRUE)
otbEdge(input=paste0(getwd(),"4490600_5321400.tif"),filter = "sobel")
}
}
\author{
Chris Reudenbach
}
