% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_calcTextures.R
\name{otbGrayMorpho}
\alias{otbGrayMorpho}
\title{Calculates Gray scale morphological operations for a given kernel size}
\usage{
otbGrayMorpho(input = NULL, out = "morpho", ram = "8192",
  filter = "dilate", structype = "ball", structype.ball.xradius = 5,
  structype.ball.yradius = 5, channel = NULL, retRaster = FALSE,
  outDir = NULL, verbose = FALSE)
}
\arguments{
\item{input}{of GeoTiff containing 1 ore more gray value bands}

\item{out}{the output mono band image containing the edge features}

\item{ram}{reserved memory in MB}

\item{filter}{the choice of the morphological operation (dilate/erode/opening/closing) (default value is dilate)}

\item{structype}{the choice of the structuring element type (ball/cross)}

\item{structype.ball.xradius}{x the ball structuring element X Radius (only if structype==ball)}

\item{structype.ball.yradius}{y the ball structuring element Y Radius (only if structype==ball)}

\item{channel}{sequence of bands to be processed}

\item{retRaster}{boolean if TRUE a raster stack is returned}

\item{verbose}{switch for system messages default is FALSE}
}
\value{
list of geotiffs containing thelocal statistics for each channel
}
\description{
Calculates Gray scale morphological operations for a given kernel size
}
\note{
the otb is used for filtering. please provide a GeoTiff file
}
\examples{
#' \dontrun{
url<-"http://www.ldbv.bayern.de/file/zip/5619/DOP\%2040_CIR.zip"
res <- curl::curl_download(url, "testdata.zip")
unzip(res,junkpaths = TRUE,overwrite = TRUE)
gm<-otbGrayMorpho(input=paste0(getwd(),"4490600_5321400.tif"),retRaster = TRUE)
raster::plot(gm[[1]])
}
}
\author{
Chris Reudenbach
}
