% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_InTex.R
\name{otbLocalStat}
\alias{otbLocalStat}
\title{Calculates local statistics for a given kernel size}
\usage{
otbLocalStat(input = NULL, out = "localStat", ram = "8192", radius = 3,
  channel = NULL, retRaster = FALSE, outDir = NULL, verbose = FALSE,
  giLinks = NULL)
}
\arguments{
\item{input}{of GeoTiff containing 1 ore more gray value bands}

\item{out}{string pattern vor individual naming of the output file(s)}

\item{ram}{reserved memory in MB}

\item{radius}{computational window in pixel}

\item{channel}{sequence of bands to be processed}

\item{retRaster}{boolean if TRUE a raster stack is returned}

\item{outDir}{output Directory}

\item{verbose}{switch for system messages default is FALSE}

\item{giLinks}{list. of GI tools cli pathes}
}
\description{
Calculates local statistics for a given kernel size
}
\note{
the otb is used for the calculation of the statistics. Please provide a GeoTiff file
}
\examples{
\dontrun{
url<-"http://www.ldbv.bayern.de/file/zip/5619/DOP\%2040_CIR.zip"
res <- curl::curl_download(url, "testdata.zip")
unzip(res,junkpaths = TRUE,overwrite = TRUE)
otblocalStat(input=paste0(getwd(),"4490600_5321400.tif"),radius=5)
}
}
\author{
Chris Reudenbach
}
