% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_pc2D_dsm.R
\name{pc2D_dsm}
\alias{pc2D_dsm}
\title{Create a Digital Terrain Model from UAV generated point clouds by minimum altitude sampling}
\usage{
pc2D_dsm(laspcFile = NULL, gisdbasePath = NULL, sampleMethod = "max",
  threshold = 20, cutExtent = NULL, targetGridSize = 0.25,
  projFolder = NULL,
  proj4 = "+proj=utm +zone=32 +datum=WGS84 +units=m +no_defs",
  giLinks = NULL, verbose = FALSE)
}
\arguments{
\item{laspcFile}{character. default is \code{NULL} path  to the laz/las file(s)}

\item{gisdbasePath}{character. default is \code{NULL} root directory of the project. NOTE the function creates two subfolder named \code{run} and \code{output}}

\item{sampleMethod}{sampling method for point aggregation}

\item{threshold}{numeric. percentile threshold}

\item{cutExtent}{clip area}

\item{targetGridSize}{numeric. the resolution of the target DTM raster}

\item{projFolder}{subfolders that will be created/linked for R related GRASS processing}

\item{proj4}{character. valid proj4 string that should be assumingly the correct one}

\item{giLinks}{list of link2GI cli pathes, default is NULL}

\item{verbose}{to be quiet (1)}
}
\description{
Create a Digital Surface Model from a high density point cloud as typically derived by an
optical UAV retrieval. It simply samples the maximum or whatever values of a given target 
grid size and fills the no data holes if so.
}
\examples{
\dontrun{

require(uavRst)
require(raster)
require(link2GI)

# proj subfolders
projRootDir<-getwd()
#setwd(paste0(projRootDir,"run"))

paths<-link2GI::initProj(projRootDir = projRootDir,
                         projFolders = c("data/","data/ref/","output/","run/","las/"),
                         global = TRUE,
                         path_prefix = "path_")

# get some colors
pal = mapview::mapviewPalette("mapviewTopoColors")

# get the data
url <- "https://github.com/gisma/gismaData/raw/master/uavRst/lidar_477375_00_5631900_00_477475_00_5632000_00.las"
res <- curl::curl_download(url, "run/lasdata.las")
# make the folders and linkages
giLinks<-uavRst::get_gi()

# create 2D pointcloud DSM
dsm <- pc2D_dsm(laspcFile = paste0(path_run,"lasdata.las"),
                gisdbasePath = projRootDir,
                sampleMethod = "max",
                targetGridSize = 0.5, 
                giLinks = giLinks)
}
}
\author{
Chris Reudenbach
}
