% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDSM.R
\name{pc2DSM}
\alias{pc2DSM}
\title{calcualte and post-classifies the morphological structure of raw tree crowns}
\usage{
pc2DSM(runDir,currentP,allP)
}
\arguments{
\item{lasDir}{default is \code{NULL} path  to the laz/las file(s)}

\item{gisdbase_path}{gisdbase will be linked or created depending on \code{gisdbase_exist}}

\item{GRASSlocation}{location will be linked or created depending on \code{gisdbase_exist}}

\item{projFolder}{subfolders in gisdbase for R related processing}

\item{grid_size}{resolution for raster operations}

\item{saga_spline_level_max}{default is 9 number ob spline iterations}

\item{otb_gauss_radius}{radius of otb smoothing filter}

\item{type_smooth}{default is \code{otb_gauss} alternatives is \code{saga_spline}}

\item{proj4}{default is EPSG 32632 any valid proj4 string that is assumingly the correct one}

\item{gisdbase_exist}{switch if gisdbase is created or  linked only}
}
\value{
pc2DSMbasically returns a  DSM
}
\description{
calcualte and post-classifies the morphological structure of raw tree crowns
}
\author{
Chris Reudenbach
}
