% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_karim.R
\name{poly_extract_maxpos}
\alias{poly_extract_maxpos}
\title{extract for all polygons the position of the maximum value of the applied raster(s)}
\usage{
poly_extract_maxpos(x, lN, poly_split = TRUE)
}
\arguments{
\item{x}{path and name of a GDAL raster file}

\item{lN}{layer name of shape file}

\item{poly_split}{split polygon in single file default is TRUE
extract for all polygons the position of the maximum value}
}
\description{
extract for all polygons the position of the maximum value
}
