% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_bluedjinn.R
\name{poly_maxpos}
\alias{poly_maxpos}
\title{extract for all polygons the position of the maximum value of the applied raster(s).}
\usage{
poly_maxpos(fileName, layerName, polySplit = TRUE, cores = 1)
}
\arguments{
\item{fileName}{path and name of a GDAL raster file}

\item{layerName}{layer name of shape file}

\item{polySplit}{split polygon in single file, default is TRUE
extract for all polygons the position of the maximum value}

\item{cores}{number of cores used}
}
\value{
returns a list of coordinates of the position of the maximum value and geometric centre of any corresponding raster object inside a given polygon
}
\description{
extract for all polygons the position of the maximum value
}
\keyword{internal}
