% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_bluedjinn.R
\name{poly_stat}
\alias{poly_stat}
\title{#TOFIX @name  @title calculate decriptive stats of raster values of underlying a polygon}
\usage{
poly_stat(x = NULL, spdf = NULL, count = 1, min = 1, max = 1,
  sum = 1, range = 1, mean = 1, var = 1, stddev = 1, quantile = 10)
}
\arguments{
\item{x}{spatial raster object}

\item{spdf}{spatial point dataframe}

\item{count}{0 1 switch}

\item{min}{0 1 switch}

\item{max}{0 1 switch}

\item{sum}{0 1 switch}

\item{range}{0 1 switch}

\item{mean}{0 1 switch}

\item{var}{0 1 switch}

\item{stddev}{0 1 switch}

\item{quantile}{number of quantile}
}
\description{
calculate statitiscs of polygon based raster extraction.Returns a spatialpolygon dataframe containing decriptive statistics
}
\examples{
\dontrun{
# Tree segementation based on a CHM
polyStat <- poly_stat(c("chm","dah"),
                      spdf = "tree_crowns.shp")
}

}
\author{
Chris Reudenbach
}
