% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_crownMetrics.R
\name{polyMetrics}
\alias{polyMetrics}
\title{calculate morphometric features of polygons}
\usage{
polyMetrics(crownarea, funNames = c("length", "elongation",
  "eccentricityboundingbox", "solidity", "eccentricityeigen", "calliper",
  "rectangularity", "circularityharalick", "convexity"))
}
\arguments{
\item{crownarea}{sp polygon object}

\item{funNames}{("length","elongation","eccentricityboundingbox","solidity","eccentricityeigen","calliper","rectangularity","circularityharalick","convexity")}
}
\description{
calculate morphometric features of polygons. calculate some crown related metrics return the metrics as a spatialpointdataframe/spatialpolygondataframe
}
\seealso{
\href{https://CRAN.R-project.org/package=Momocs}{Momocs}
\href{https://www.researchgate.net/profile/Paul_Rosin/publication/228382248_Computing_global_shape_measures/links/0fcfd510802e598c31000000.pdf?origin=publication_detail}{Paul Rosin}
}
