% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_classify.R
\name{predict_rgb}
\alias{predict_rgb}
\title{classify images using raster predict}
\usage{
predict_rgb(imageFiles = NULL, model = NULL, inPrefix = "index_",
  outPrefix = "classified_", bandNames = NULL)
}
\arguments{
\item{imageFiles}{raster*. imagestack for classification purposes must contain the required bands as needed by the model.}

\item{model}{model. classification model}

\item{inPrefix}{character. in frefix  string}

\item{outPrefix}{character. out prefix string}

\item{bandNames}{character. band names}
}
\description{
classify images using raster predict
}
\examples{
\dontrun{
predict_rgb(imageFiles=imagestack,
            model = model_final,
            inPrefix = "index_",
            outPrefix = "classified_",
            bandNames = c("R","G","B","A","VARI","NDTI","TGI","GLI","NGRDI"))
}
}
