% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_classify.R
\name{predictRGB}
\alias{predictRGB}
\title{classify images using raster predict}
\usage{
predictRGB(imageFiles = NULL, model = NULL, in_prefix = "index_",
  out_prefix = "classified_", bandNames = c("R", "G", "B", "A", "VARI",
  "NDTI", "TGI", "GLI", "NGRDI", "GLAI", "Mean", "Variance", "Skewness",
  "Kurtosis"))
}
\arguments{
\item{imageFiles}{imagestacke for classification}

\item{model}{classification model}

\item{in_prefix}{in frefix  string}

\item{out_prefix}{out prefix string}

\item{bandNames}{band names}
}
\description{
classify images using raster predict
}
