% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_karim.R
\name{R2SAGA}
\alias{R2SAGA}
\title{converts SAGA raster to R raster object}
\usage{
R2SAGA(x, fn)
}
\arguments{
\item{x}{raster object}

\item{fn}{filname without extension}
}
\description{
converts SAGA raster to R raster object
}
