% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_bluedjinn.R
\name{raster2sdat}
\alias{raster2sdat}
\title{writes raster* objects to the SAGA format}
\usage{
raster2sdat(rastobj, path_run = tempdir())
}
\arguments{
\item{rastobj}{raster* object}

\item{path_run}{character. path used for runtime operations}
}
\description{
writes raster* objects to the SAGA format
}
\keyword{internal}
