% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCost.R
\name{rcost}
\alias{rcost}
\title{Calculate the least cost path from point to point using a given friction matrix}
\usage{
rcost(pointList,costraster)
}
\arguments{
\item{pointList}{containing coords}

\item{a}{cost raster object}
}
\value{
rcost returns matrix cost/transformed real distances
}
\description{
http://moc.environmentalinformatics-marburg.de/doku.php?id=courses:msc:advanced-gis:description
}
\examples{
#### calculate least cost pathes

      costMatrix<-rcostance(pointList,dem)
}
\author{
Chris Reudenbach
}

