% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcShed.R
\name{rcShed}
\alias{rcShed}
\title{calculates the viewshed alias remote control range for a given launching point and the calculated flight altitude.}
\usage{
rcShed(runDir,launchP,flightAlt)
}
\arguments{
\item{launchP}{actual launching position aka position of the RC}

\item{flightAlt}{flight altitude}

\item{runDir}{the choosen working directory}
}
\value{
rcShed basically returns mask raster 0= no/poor RC signal 1=sane RC signal
}
\description{
viewshed analysis to derive an rough estimation of rc range.
}
\examples{
#### Example to use rcShed for a typical uav flight
# setup GIS environ
envGIS<- initRGIS("~/proj/drone",'uniwald',"~/proj/drone/uniwald/mrbiko.tif")  
# call rcShed
rcRange<-rcShed(envGIS,launchP = c(8.692, 50.842316), flightAlt = 100, rcRange = 1000,dem="mrbiko.tif")

}
\author{
Chris Reudenbach
}

