% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_bluedjinn.R
\name{read_gpx}
\alias{read_gpx}
\title{Read a GPX file.}
\usage{
read_gpx(
  file,
  layers = c("waypoints", "tracks", "routes", "track_points", "route_points")
)
}
\arguments{
\item{file}{a GPX filename (including directory)}

\item{layers}{vector of GPX layers. Possible options are \code{"waypoints"}, \code{"tracks"}, \code{"routes"}, \code{"track_points"}, \code{"route_points"}. By dedault, all those layers are read.}
}
\value{
a sp* object containing the gpx track objects
}
\description{
Read a GPX file. By default, it reads all possible GPX layers, and only returns shapes for layers that have any features.
if the layer has any features a sp object is returned.
}
\note{
adapted from \code{\link[tmaptools]{read_GPX}}
}
\keyword{internal}
