% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusionTools.R
\name{readMetrics}
\alias{readMetrics}
\alias{GridMetrics}
\title{lists only the desired densityMetrics asc from outputfolder}
\usage{
GridMetrics(lasFiles, res, heightClassList, metrics, heightbreak, buffer,
  cellbuffer, strata, instrata)
}
\arguments{
\item{lasFiles}{default is \code{NULL} list of the las file(s)}

\item{res}{default is \code{NULL} cellsize}

\item{metrics}{default is  \code{list(c(0,2,5,10,30))} list of metric indexes indexes given
in Fusion Manual p.78/79}

\item{heightbreak}{lower heightbreak}

\item{buffer}{optional parameter \code{NULL} if extent should be extendet by a buffer}

\item{cellbuffer}{optional parameter \code{NULL} if cellbased metrics should
take neighbouring cells into account}

\item{strata}{optional paramter \code{list(c(0,2,5,10,30))}, elevation based metrics
will be calculated for strata}

\item{instrata}{optional paramter \code{list(c(0,2,5,10,30))}, intensity based metrics
will be calculated for strata}

\item{heightClasses}{default is \code{list(c(0,2,5,10,30))}}

\item{lasFiles}{default is \code{NULL} list of the las file(s)}

\item{heightClasses}{default is \code{list(c(0,2,5,10,30))}}
}
\description{
lists only the desired densityMetrics asc from outputfolder

calculate elevation and intensity based metrics an write out selected metrics to raster (.asc)
}
\examples{
\dontrun{
 GridList<-readAscGridMetrics(lasFiles = lasFiles, heightClasses = heights)

}


\dontrun{

GridMetrics(lasFiles = lasFiles, res = gridsize, heightClassList = h, metrics = 31, heightbreak=0.2 )
}


}
\author{
Chris Reudenbach,Jannis Gottwald

Chris Reudenbach,Jannis Gottwald
}
