% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readxyz.R
\name{readxyz}
\alias{readxyz}
\title{Read and Convert xyz DEM/DSM Data as typically provided by the Authorities}
\usage{
readxyz(xyzFN = NUL, epsgCode = "25832")
}
\arguments{
\item{txtFn}{ASCII tect file with xyz values}
}
\value{
A \code{RasterLayer}
}
\description{
Read xyz data and generate a raster  \code{Raster*} object.
}
\examples{

# get some typical authority generated data
url<-"http://www.ldbv.bayern.de/file/zip/10430/DGM_1_ascii.zip"
res <- curl::curl_download(url, "testdata.zip")
unzip(res,files = grep(".tif", unzip(res,list = TRUE)$Name,value = TRUE),junkpaths = TRUE,overwrite = TRUE)
readxyz(file.path(getwd(),basename(grep(".tif", unzip(res,list = TRUE)$Name,value = TRUE))))

}
\author{
Chris Reudenbach
}

