% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robubu-package.R
\docType{package}
\name{robubu-package}
\alias{robubu-package}
\title{the rolling burning bus - awesome gis tools fyi}
\description{
the rolling burning bus - whatever, it runs fine for now
}
\details{
The package provides some hybrid R-GI functionality

The package name is a tribute to Randall Munroe's great xkcd webcomics and especially \href{http://xkcd.com/1695/}{Code Quality 2}
}
\author{
Chris Reudenbach,
\cr
\emph{Maintainer:} Chris Reudenbach \email{reudenbach@uni-marburg.de}
}
\keyword{package}

