% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBeetle.R
\name{runBeetle}
\alias{runBeetle}
\title{Wrapper function to analyse the phylogenetic spread of running beetle
 via a cost analysis based on morphometry and derived potential surface wetness.}
\usage{
runBeetle(rootDir,workingDir="cost",inputData=NULL,costType="tci",
 externalCostRaster=NULL,internalCost=TRUE)
}
\arguments{
\item{rootDir}{project directory\cr}

\item{workingDir}{working directory\cr}

\item{inputData}{location data containing obligatory the cols lon,lat and
optional a code col. The format has to be a data frame see example.}

\item{costType}{used if internalCost = TRUE.  default is "tci"  you can choose
"tci" "dem.filled" or "accu" see details for more information}

\item{internalCost}{default = TRUE switches to external provided GTiff file
wich has to be named cost.tifx}

\item{dump}{default = FALSE if TRUE export r.terraflow products to GTiff}

\item{usedump}{default = FALSE instead of running  r.terraflow again use
products products to GTiff}

\item{externalCostRaster}{default = NULL you may provide a GTiff file with the
name cost.tif NOTE you have to set internalCost = FALSE}
}
\value{
runBeetle returns:\cr
(a) a dataframe with the (a) euclidian distances, (b) cost distance (isotropic cost surface) and (c) the walk distance (anisotropic cost surface)\cr
}
\description{
The phylogenetic spreading of the tibetan carabidae seems to be
correlated to valley linkage and available humidity. Long term humidity is
bound to precipitation and morphometry. It is obvious that the estimation of
spreeading speed and range is in no case a simple euclidian one. runBeetle
provides a first better estimation using a cost or friction analysis assuming
that the spread is following natural lines of wetness e.g. valleys, humidity
gradients...whatever  I'am not a beetle ask the beetle
\href{http://www.zoologie.uni-rostock.de/mitarbeiter/joachimschmidt/joachimschmidtpubl}{beetle
guy}.
}
\details{
The core of the analysis is an isotropic/anisotropic least cost path
 calculation. By default the cost surface is assumed to be a local derivate
 of the morphometry with respect to the potential soil humidity. A perfect
 approch to derive such information is the use of a Digital Elevation Model
 DEM and some corresponding derivates as the Topographic Convergence Index.\cr

 If you choose "tci"  (default) the cost surface provides an estimation of
 rainwater runoff availability to plants based on specific catchment area (A)
 and local slope (b) such that TCI = ln(A/tan b) (Beven & Kirkby, 1979). This
 seems be pretty straightforward and fairly suitable for the beetles
 "behaviour.\cr

 If you chosse "accu" a typical accumulation cost grid from the original DEM
 will be used\cr

 If you choose "dem.filled" a typical accumulation cost grid from the
 hydrologically corrected DEM will be used\cr

 For the r.walk algorithm the non accumulated data is used.
}
\examples{

\dontrun{
#### NOTE: You obligatory need GRASS70

library(doParallel)
library(foreach)
library(raster)
library(sp)
library(gdalUtils)
library(rgdal)

### NOTE: the area is specifified and automatically downloaded (SRTM) by the input data extent plus a "zone"
### read a csv file
fn<-system.file("bug.csv", package="biomix")
beetleLocs<-read.csv2(fn,header = TRUE,sep = ',',dec = '.',stringsAsFactors=FALSE)

### use some arbitrary locations from scratch
beetleLocs <- as.list(c("86.83", "28.20", "100","84.58", "28.67", "200" ,"83.87", "28.80", "300"))
beetleLocs <- data.frame(matrix(as.numeric(unlist(beetleLocs)), nrow=3, byrow=T),stringsAsFactors=FALSE)
colnames(beetleLocs)<-c("lon","lat","code")
beetleDist<-runBeetle(rootDir = "/home/creu/proj/beetle" ,inputData = beetleLocs, usedump=TRUE, walk=TRUE)
}
}
\author{
Chris Reudenbach
\cr
\emph{Maintainer:} Chris REudenbach \email{reudenbach@uni-marburg.de}
}
\references{
Schmidt, J., Böhner, J., Brandl, R. & Opgenoorth, L. (in review):
 Mass elevation and lee effect override latitudinal effects in determining
 the distribution ranges of species: Ground beetles from the Himalaya-Tibet
 Orogen. – PLoS ONE.

\href{https://grass.osgeo.org/grass7/}{GRASS70}
\href{https://sourceforge.net/projects/saga-gis/files/}{SAGA GIS}
}

