% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_bluedjinn.R
\name{searchLastools}
\alias{searchLastools}
\title{#' clips a tif files according to a given extent.
#'
#' clips a tif files according to a given extent
#' @param rasterFiles character. vector containing a list of rasterfiles to be clipped
#' @param ext extent
#' @param outPath character. subfolder of current runtime folder. clipped files will be stored there
#' @param prefix character. prefic string that is added to the filenames
#'@export
#'@keywords internal
cutTif<- function(rasterFiles = NULL,
                  ext=NULL,
                  outPath="cut",
                  prefix="cut") {
  #rasterFiles <- list.files(pattern="[.]tif$", path="/home/creu/proj/geopat/data/modis_carpathian_mountains/study_area/modis_ndvi/2002", full.names=TRUE)
  te=paste(raster::extent(ext)[1],' ',
           raster::extent(ext)[3],' ',
           raster::extent(ext)[2],' ',
           raster::extent(ext)[4])
  if (!file.exists(file.path(R.utils::getAbsolutePath(path_run), outPath))) dir.create(file.path(file.path(R.utils::getAbsolutePath(path_run), outPath)), recursive = TRUE,showWarnings = FALSE)
  for (rasterFile in rasterFiles) {
    system(paste0("gdal_translate -projwin ", te, " -of GTiff ",rasterFile, " ", file.path(R.utils::getAbsolutePath(path_run)),outPath,"/",prefix,basename(rasterFile)))
  }
}4
search for lastools}
\usage{
searchLastools(MP = "~", quiet = TRUE)
}
\arguments{
\item{MP}{mount point point for starting search}

\item{quiet}{default = TRUE}
}
\description{
search for lastools
#'@keywords internal
}
