% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_leafSel.R
\name{select}
\alias{select}
\title{selecting vector features using leaflet draw}
\usage{
select(mapCenter = c(50.80801, 8.72993), zoom = 15, position = "topright",
  maplayer = c("CartoDB.Positron", "OpenStreetMap", "Esri.WorldImagery",
  "Thunderforest.Landscape", "OpenTopoMap"), overlay = NULL, cex = 10,
  lwd = 2, alpha = 0.6, opacity = 0.7)
}
\arguments{
\item{zoom}{initial zoom level}

\item{position}{place to put the toolbar (topright, topleft, bottomright, bottomleft)}

\item{maplayer}{string as provided by leaflet-provider}

\item{overlay}{optional sp object}

\item{mapcenter}{c(lat,lon) central point of the leaflet map}

\item{line}{enable the draw tool line tool}

\item{poly}{enable the draw polygon tool}

\item{circle}{enable the draw circle tool}

\item{point}{enable the draw point tool}

\item{remove}{enable/disable the remove feature of the draw tool}

\item{intersection}{enable/disable th possibility to overlay lines or polygons}

\item{preset}{textstring "NULL" full draw version, "uav" for flightarea digitizing, "ext" for rectangles}
}
\description{
select is based on the leaflet draw plugin. It provides a bunch of leaflet maps as base layers for digitizing vector features.
}
\note{
Yu can either save the digitized object to a json file or you ma grab the json string via the clipboard
}
\examples{

# all features
select()

# preset for digitizing uav flight areas
select(preset="uav")

#' # preset for digitizing extents
select(preset="ext")

}
\author{
Chris Reudenbach
}
