% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_bluedjinn.R
\name{setHomePath}
\alias{setHomePath}
\title{Checks if running on a specified computer domain}
\usage{
setHomePath(homeDir = "F:/MPG", prefixPC = "PCRZP")
}
\arguments{
\item{homeDir}{full path  to the real folder location the project}

\item{prefixPC}{contains  an arbitrary part of the computer name. It always starts with the first letter.}
}
\description{
Checks if the computer name belongs to a specified group i.e. aq network domain Marburg Universitys computer pools
}
\examples{
\dontrun{
# add path
setHomePath("saga",prefixPC="PCRZP")
}
}
\author{
CR
}
\keyword{internal}
