% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_bluedjinn.R
\name{shp2so}
\alias{shp2so}
\title{reads shapefiles as sf or sp objects}
\usage{
shp2so(fn, type = "sp", path_run = tempdir())
}
\arguments{
\item{fn}{filename optionally with path}

\item{type}{character. type of object to be created. "sp" is default can be set to "sf"}

\item{path_run}{character. path used for runtime operations}
}
\description{
reads shapefiles as sf or sp objects
}
\keyword{internal}
