% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_bluedjinn.R
\name{so2shp}
\alias{so2shp}
\title{writes shapefiles from  sf or sp* objects}
\usage{
so2shp(spobj, name = "spobj", path_run = tempdir())
}
\arguments{
\item{spobj}{spatial object of type sp* or sf}

\item{name}{character. name of object to be created}

\item{path_run}{character. path used for runtime operations}
}
\description{
writes shapefiles from  sf or sp* objects
}
\keyword{internal}
