% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSoloLog.R
\name{solo2gpx}
\alias{solo2gpx}
\title{download tlog files from solo and convert them to gpx files}
\usage{
solo2gpx(logFiles = "solo.t*", logDir = "soloLog", download = TRUE,
  netWarn = TRUE, import = FALSE)
}
\arguments{
\item{logFiles}{pattern of which kind of logs should be downloaded for telemetry it is "solo.t*"}

\item{logDir}{path to the folder where the logs should be downloaded to}

\item{download}{boolean if you want to download log files from the solo controller}

\item{netWarn}{if true warns and waits before starting a connection to the controller to connect to the solo wifi}

\item{import}{import of the selecected as sp objects}
}
\description{
solo2gpx is interfacing the mavtogpx.py converter
}
\note{
for using the solo stuff you need to install: sudo pip install pymavlink; sudo pip install dronekit-sitl; sudo pip install dronekit; sudo apt-get install sshpass
}
\examples{

## download current telemetry log file from controller and convert it to gpx
solo2gpx(logDir="~/tmp/solo",logFiles = "solo.tlog")

## download all available telemetry logfiles from the controller
solo2gpx(logDir="~/tmp/solo")

## download ALL logfiles from the controller
solo2gpx(logDir="~/tmp/solo", logFiles = "*")

}
\author{
Chris Reudenbach
}

