% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_solo.R
\name{soloLog}
\alias{soloLog}
\title{download tlog files from solo and convert them to gpx files}
\usage{
soloLog(logFiles = "solo.t*", logDir = "~/soloLog", downloadOnly = FALSE,
  netWarn = TRUE, organize = TRUE, makeSP = FALSE)
}
\arguments{
\item{logFiles}{pattern of which kind of logs should be downloaded for telemetry it is "solo.t*"}

\item{logDir}{path to the folder where the logs should be downloaded to}

\item{downloadOnly}{boolean if you ONLY want to download the log files from the solo controller to a directory}

\item{netWarn}{if true warns and waits before starting a connection to the controller to connect to the solo wifi}

\item{organize}{renames the log and gpx files according to their timeperiod}

\item{makeSP}{generates SP objects from the gpx files}
}
\description{
soloLogs is interfacing the mavtogpx.py converter. It downloads and/ or converts the 3DR Solo logfiles
}
\note{
for using the solo stuff you need to install: sudo pip install pymavlink; sudo pip install dronekit-sitl; sudo pip install dronekit; sudo apt-get install sshpass
}
\examples{

## download current telemetry log file from controller and convert it to gpx
soloLog(logDir="~/tmp/solo",logFiles = "solo.tlog")

## download all available telemetry logfiles from the controller
soloLog(logDir="~/tmp/solo")

## download ALL logfiles from the controller
soloLog(logDir="~/tmp/solo", logFiles = "*")

}
\author{
Chris Reudenbach
}
