% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_karim.R
\name{sp_line}
\alias{sp_line}
\title{create an spatiallineobject from 2 points}
\usage{
sp_line(p1, p2, ID, proj4 = "+proj=longlat +datum=WGS84 +no_defs",
  export = FALSE)
}
\arguments{
\item{p1}{coordinate of first point}

\item{p2}{coordinate of second point}

\item{ID}{id of line}

\item{proj4}{proj4 string}

\item{export}{write shafefile default = F}
}
\description{
create an spatiallineobject from 2 points, optional export as shapefile
}
