% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_bluedjinn.R
\name{sp_point}
\alias{sp_point}
\title{create an spatialpointobject from 1 points.}
\usage{
sp_point(lon, lat, ID = "point",
  proj4 = "+proj=longlat +datum=WGS84 +no_defs", export = FALSE)
}
\arguments{
\item{lon}{numeric. .i longitude of point}

\item{lat}{numeric. ie latitude of point}

\item{ID}{name of point}

\item{proj4}{proj4 string}

\item{export}{write shafefile default = F}
}
\description{
create an spatialpointobject from 1 points, optional export as shapefile
#@author Chris Reudenbach
}
\keyword{internal}
