% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_calcTextures.R, R/tb_calcTextureChannels.R
\name{textureVariables}
\alias{textureVariables}
\alias{textureVariables}
\title{This functions calls the glcm package with standard settings}
\usage{
textureVariables(x, nrasters = 1:nlayers(x), kernelSize = c(3),
  stats = c("mean", "variance", "homogeneity", "contrast", "dissimilarity",
  "entropy", "second_moment", "correlation"), shift = list(c(0, 1), c(1, 1),
  c(1, 0), c(1, -1)), parallel = TRUE, n_grey = 8, min_x = NULL,
  max_x = NULL)

textureVariables(x, nrasters = 1:nlayers(x), kernelSize = c(3),
  stats = c("mean", "variance", "homogeneity", "contrast", "dissimilarity",
  "entropy", "second_moment", "correlation"), shift = list(c(0, 1), c(1, 1),
  c(1, 0), c(1, -1)), parallel = TRUE, n_grey = 8, min_x = NULL,
  max_x = NULL)
}
\arguments{
\item{x}{rasterLayer or a rasterStack containing different channels}

\item{nrasters}{vector of channels to use from x. Default =nlayers(x)}

\item{kernelSize}{vector of numbers indicating the environment sizes for which the textures are calculated}

\item{stats}{string vector of parameters to be calculated.see \code{\link{glcm}}}

\item{parallel}{logical value indicating whether parameters are calculated parallel or not}

\item{n_grey}{number of grey values. see \code{\link{glcm}}}

\item{min_x}{for each channel the minimum value which can occur. If NULL then the minimum value from the rasterLayer is used.}

\item{max_x}{for each channel the maximum value which can occur. If NULL then the maximum value from the rasterLayer is used.
This functions calls the glcm function from \link{glcm} with standard settings
and returns list of RasterStacks containing the texture parameters for each combination of channel and kernelSize}

\item{x}{rasterLayer or a rasterStack containing different channels}

\item{nrasters}{vector of channels to use from x. Default =nlayers(x)}

\item{kernelSize}{vector of numbers indicating the environment sizes for which the textures are calculated}

\item{stats}{string vector of parameters to be calculated.see \code{\link{glcm}}}

\item{n_grey}{number of grey values. see \code{\link{glcm}}}

\item{parallel}{logical value indicating whether parameters are calculated parallel or not}

\item{min_x}{for each channel the minimum value which can occur. If NULL then the minimum value from the rasterLayer is used.}

\item{max_x}{for each channel the maximum value which can occur. If NULL then the maximum value from the rasterLayer is used.
This functions calls the glcm function from \link{glcm} with standard settings
and returns list of RasterStacks containing the texture parameters for each combination of channel and kernelSize}
}
\description{
This functions calls the glcm package with standard settings

This functions calls the glcm package with standard settings
}
\note{
for the use of textureVariables a glcm wrapper function 
      a raster* object is required

More information at the texture tutorial site of
\link{http://www.fp.ucalgary.ca/mhallbey/more_informaton.htm}(Mryka Hall-Beyer)
Keep in mind that:\cr
Homogeneity is correlated with Contrast,  r = -0.80
Homogeneity is correlated with Dissimilarity, r = -0.95
GLCM Variance is correlated with Contrast,  r= 0.89
GLCM Variance is correlated with Dissimilarity,  r= 0.91
GLCM Variance is correlated with Homogeneity,  r= -0.83
Entropy is correlated with ASM,  r= -0.87
GLCM Mean and Correlation are more independent. For the same image,
GLCM Mean shows  r< 0.1 with any of the other texture measures demonstrated in this tutorial.
GLCM Correlation shows  r<0.5 with any other measure.

for the use of textureVariables a glcm wrapper function 
      a raster* object is required

More information at the texture tutorial site of
\link{http://www.fp.ucalgary.ca/mhallbey/more_informaton.htm}(Mryka Hall-Beyer)
Keep in mind that:\cr
Homogeneity is correlated with Contrast,  r = -0.80
Homogeneity is correlated with Dissimilarity, r = -0.95
GLCM Variance is correlated with Contrast,  r= 0.89
GLCM Variance is correlated with Dissimilarity,  r= 0.91
GLCM Variance is correlated with Homogeneity,  r= -0.83
Entropy is correlated with ASM,  r= -0.87
GLCM Mean and Correlation are more independent. For the same image,
GLCM Mean shows  r< 0.1 with any of the other texture measures demonstrated in this tutorial.
GLCM Correlation shows  r<0.5 with any other measure.
}
\examples{
#' \dontrun{
## example on how to calculate texture from a list of channels

url<-"http://www.ldbv.bayern.de/file/zip/5619/DOP\%2040_CIR.zip"
res <- curl::curl_download(url, "testdata.zip")
unzip(res,junkpaths = TRUE,overwrite = TRUE)
r<- raster::stack(paste0(getwd(),"4490600_5321400.tif"))

# call glcm wrapper
result <- textureVariables(r,nrasters=1:3,
stats=c("mean", "variance", "homogeneity"))

#plot the results from VIS0.6 channel:
raster::plot(unlist(unlist(glcm$size_3$X4490600_5321400.1)))
}
#' \dontrun{
## example on how to calculate texture from a list of channels

url<-"http://www.ldbv.bayern.de/file/zip/5619/DOP\%2040_CIR.zip"
res <- curl::curl_download(url, "testdata.zip")
unzip(res,junkpaths = TRUE,overwrite = TRUE)
r<- raster::stack(paste0(getwd(),"4490600_5321400.tif"))

# call glcm wrapper
result <- textureVariables(r,nrasters=1:3,
stats=c("mean", "variance", "homogeneity"))

#plot the results from VIS0.6 channel:
raster::plot(unlist(unlist(glcm$size_3$X4490600_5321400.1)))
}
}
\seealso{
\code{\link{glcm}}

\code{\link{glcm}}
}
\author{
Hanna Meyer

Hanna Meyer
}
