% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_karim.R
\name{Tiff2G}
\alias{Tiff2G}
\title{converts geotiff to GRASS raster}
\usage{
Tiff2G(runDir = NULL, layer = NULL)
}
\arguments{
\item{runDir}{path of working directory}

\item{layer}{name GRASS raster}
}
\description{
converts geotiff to GRASS raster
}
