% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_lidr_makros.R
\name{tile_eq}
\alias{tile_eq}
\title{divide a lidR las* or lascatalog object by given factors}
\usage{
tile_eq(cgs, factorx = 4, factory = 4, outpath = NULL,
  lasfn = "cut",
  proj4 = "+proj=utm +zone=32 +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0")
}
\arguments{
\item{cgs}{filename of an las file}

\item{factorx}{factor to divide in x direction}

\item{factory}{factor to divide in y direction}

\item{outpath}{path to write to if null nothing is written}

\item{lasfn}{filname of lasfile if written}

\item{proj4}{correct proj4 string}
}
\description{
divide a lidR las* object by given factors. The resulting tiles are  written to the outpath folder
}
\examples{
\dontrun{
cut_eq(cgs = lidR::readLAS(las_file),factorx = 4,factory = 4 )
}
}
