% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_classify.R
\name{trainModel}
\alias{trainModel}
\title{trains model according to trainingdata and image stacks}
\usage{
trainModel(trainingDF = NULL, predictors = c("R", "G", "B"),
  response = "ID", spaceVar = "FN", names = c("ID", "R", "G", "B", "A",
  "FN"), noLoc = 3, cl_method = "rf", metric_ffs = "Kappa",
  metric_caret = "ROC", pVal = 0.5, prefin = "final_", preffs = "ffs_",
  modelSaveName = "model.RData", nrclu = 3)
}
\arguments{
\item{predictors}{vector of predictor names as given by the header of the training data table}

\item{response}{name of response variable as given by the header of the training data table}

\item{spaceVar}{name of the spcetime splitting vatiable as given by the header of the training data table}

\item{names}{all names of the dataframe header}

\item{noLoc}{number of locations to leave out usually nuber of dicrete trainings locations/images}

\item{cl_method}{classification method default is \code{"rf"}}

\item{metric_ffs}{accuracy metrics for ffs for classification  default is \code{"kappa"}}

\item{metric_caret}{accuracy metrics for optimisation default is \code{ "ROC"}}

\item{pVal}{used part of the training data  default is \code{ 0.5}}

\item{prefin}{name pattern used for model default is \code{"final_"}}

\item{preffs}{name pattern used for ffs default is \code{"ffs_"}}

\item{modelSaveName}{name pattern used for saving the model default is \code{"model.RData" }}

\item{imageFiles}{imagestacke for classification}

\item{model}{classification model}

\item{in_prefix}{in frefix  string}

\item{out_prefix}{out prefix string}

\item{bandNames}{band names}
}
\description{
trains model according to trainingdata and image stacks
}
\examples{
 
#' \dontrun{
result<-  trainModel(trainingDF =trainingDF,
                     predictors   = c("R","G","B","VARI","NDTI","TGI","GLI","GLAI"),
                     response     = "ID",
                     spaceVar     = "FN",
                     names = c("ID","R","G","B","A","VARI","NDTI","TGI","GLI","GLAI","FN"),
                     noLoc        = length(imageTrainFiles),
                     cl_method    = "rf",
                     metric_ffs   = "kappa",
                     metric_caret = "ROC",
                     pVal         = 0.5) 
                 }
}
