% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_treePosition.R
\name{treePos}
\alias{treePos}
\title{Find potential tree positions using a canopy height model}
\usage{
treePos(chm = NULL, minTreeAlt = 10, minTreeAltParam = "chmQ20",
  minCrownArea = 3, maxCrownArea = 150, output = 1, join = 1,
  thresh = 0.1, split = TRUE, giLinks = NULL)
}
\arguments{
\item{chm}{spatial raster object}

\item{minTreeAlt}{default is 5}

\item{minTreeAltParam}{default is "chmQ20"}

\item{minCrownArea}{default is 3 minimum area of crown}

\item{maxCrownArea}{default is 225 maximum area of crown}

\item{output}{default is 0,     # 0=s treePos value 1=segment id}

\item{join}{default is 2,     # 0=no join, 1=treePos2saddle diff, 2=treePos2treePos diff}

\item{thresh}{default is 0.05,  # threshold for join difference in m}

\item{split}{default  is TRUE switch if splitting of the polygons is called}
}
\description{
Find potential tree positions using a canopy height model using a iterative watershed algorithm. return basically returns a  vector data sets with the tree crown geometries and a bunch of corresponding indices
}
\examples{
\dontrun{
# Tree segmentation based on a CHM
 treePos(chm = rasterobj,  "nameofSAGAFile")
}

}
\author{
Chris Reudenbach
}
