% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_treePosition.R
\name{treePoslidR}
\alias{treePoslidR}
\title{Tree top detection based on local maxima filters 'lidR'}
\usage{
treePoslidR(chm = NULL, movingWin = 7, minTreeAlt = 2)
}
\arguments{
\item{chm}{Canopy height model in \link[raster]{raster}, \code{lasmetrics}, \code{matrix} or  object of \code{class LAS}.
Should be the same that was used to create
the input for \code{treePos}.}

\item{movingWin}{Size (in pixels) of the moving window to detect local maxima.}

\item{minTreeAlt}{Height threshold (m) below a pixel cannot be a local maximum. Local maxima values are used to define tree tops.}
}
\description{
Tree top detection based on local maxima filters. There are two types of filter. The first,
called for gridded objects, works on images with a matrix-based algorithm and the second one, called for 
point clouds, works at the point cloud level without any rasterization. Jean-Romain Roussel and David Auty:
R package \href{https://CRAN.R-project.org/package=lidR}{lidR}\cr
}
\examples{
\dontrun{
 crownslidR <-treePoslidR(chm,fws,minht)
}
}
