% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_treePosition.R
\name{treePosRL}
\alias{treePosRL}
\title{individual tree detection whitin the LiDAR-derived Canopy Height Model (CHM) 'rLiDAR'}
\usage{
treePosRL(chm = NULL, movingWin = 7, minTreeAlt = 2)
}
\arguments{
\item{chm}{Canopy height model in \link[raster]{raster} or \link[raster]{SpatialGridDataFrame} file format. Should be the same that was used to create
the input for \code{treePos}.}

\item{movingWin}{Size (in pixels) of the moving window to detect local maxima.}

\item{minTreeAlt}{Height threshold (m) below a pixel cannot be a local maximum. Local maxima values are used to define tree tops.}
}
\description{
Detects and computes the location and height of individual trees within 
the LiDAR-derived Canopy Height Model (CHM). The algorithm implemented in this function 
is local maximum with a fixed window size. Carlos A. Silva et all.: R package \href{https://CRAN.R-project.org/package=rLiDAR}{rLiDAR}\cr
}
\examples{
\dontrun{
 treePosITC <- treePosRL(chm,fws,minht)
}
}
