% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_solo.R
\name{upload2Solo}
\alias{upload2Solo}
\title{upload mission file to solo}
\usage{
upload2Solo(missionFile = NULL, connection = "udp:10.1.1.166:14550",
  prearm = "-9")
}
\arguments{
\item{missionFile}{mission file to upload}

\item{connection}{a valid connection string to the Solo default is "udp:10.1.1.166:14550"}

\item{prearm}{controls the prearm status of the Solo prearm check 0=Disabled,1=Enabled,-3=Skip Baro,-5=Skip Compass,-9=Skip GPS,-17=Skip INS,-33=Skip Params/Rangefinder,-65=Skip RC,127=Skip Voltage}
}
\description{
upload2Solo provides a crude interface to upload the Solo mission file to the 3dr SOLO
}
\note{
#' @note for using the solo stuff you need to install: sudo pip install pymavlink; sudo pip install dronekit-sitl; sudo pip install dronekit; sudo apt-get install sshpass
}
\examples{




upload2Solo("export_1001_solo.waypoints")

}
