% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robubu-package.R
\docType{data}
\name{world}
\alias{world}
\title{world boundary data set}
\format{\code{"SpatialPolygonDataFrame-class"}}
\source{
\url{https://cran.r-project.org/web/packages/mapdata}
}
\description{
world boundary data set
}
\details{
This dataset is from the mapdata package
\code{\link{mapdata}}.
}

