% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_xpolystat.R
\name{xpolystat}
\alias{xpolystat}
\title{calculate statitiscs of polygon based raster extraction}
\usage{
xpolystat(x = NULL, spdf = NULL, count = 1, min = 1, max = 1,
  sum = 1, range = 1, mean = 1, var = 1, stddev = 1, quantile = 10)
}
\arguments{
\item{x}{spatial raster object}

\item{minTreealt}{default is 5}
}
\value{
returns a spatialpolygon dataframe containing decriptive statistics
}
\description{
calculate statitiscs of polygon based raster extraction
}
\examples{
\dontrun{
# Tree segementation based on a CHM
polyStat <- xpolystat(c("chm","dah"),
                      spdf = "tree_crowns.shp")
}

}
\author{
Chris Reudenbach
}
