% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_crownseg.R
\name{chmseg_ITC}
\alias{chmseg_ITC}
\title{Decision tree segmentation method to grow individual tree crowns based on 'itcSegment'}
\usage{
chmseg_ITC(chm = NULL, EPSG = 3064, movingWin = 7,
  TRESHSeed = 0.45, TRESHCrown = 0.55, minTreeAlt = 2,
  maxCrownArea = 100)
}
\arguments{
\item{chm}{raster*, Canopy height model in \code{raster} or \code{SpatialGridDataFrame} file format. Should be the same that was used to create
the input for \code{treepos}.}

\item{EPSG}{character. The EPSG code of the reference system of the CHM raster image.}

\item{movingWin}{numeric. Size (in pixels) of the moving window to detect local maxima. \href{https://CRAN.R-project.org/package=itcSegment}{itcSegment}}

\item{TRESHSeed}{numeric. seeding threshold. \href{https://CRAN.R-project.org/package=itcSegment}{itcSegment}}

\item{TRESHCrown}{numeric. crowns threshold. \href{https://CRAN.R-project.org/package=itcSegment}{itcSegment}}

\item{minTreeAlt}{numeric. Height threshold (m) below a pixel cannot be a local maximum. Local maxima values are used to define tree tops.\href{https://CRAN.R-project.org/package=itcSegment}{itcSegment}}

\item{maxCrownArea}{numeric. A single value of the maximum individual tree crown radius expected. Default 10.0 m.
height of \code{treepos}.}
}
\description{
Segmentation of individual tree crowns as polygons based on a LiDAR derived canopy height model.
Michele Dalponte: R package \href{https://CRAN.R-project.org/package=itcSegment}{itcSegment}.
 M. Dalponte, F. Reyes, K. Kandare, and D. Gianelle,
 "Delineation of Individual Tree Crowns from ALS and Hyperspectral data: a comparison among four methods,"
 European Journal of Remote Sensing, Vol. 48, pp. 365-382, 2015.
}
\examples{
\dontrun{

##- required packages
require(uavRst)
require(link2GI)

##- create and check the links to the GI software
giLinks<-uavRst::linkAll()
if (giLinks$saga$exist & giLinks$otb$exist & giLinks$grass$exist) {

##- project folder
projRootDir<-tempdir()

##- create subfolders please mind that the pathes are exported as global variables
paths<-link2GI::initProj(projRootDir = projRootDir,
                         projFolders = c("data/","data/ref/","output/","run/","las/"),
                         global = TRUE,
                         path_prefix = "path_")
##- overide trailing backslash issue
path_run<-ifelse(Sys.info()["sysname"]=="Windows", sub("/$", "",path_run),path_run)
setwd(path_run)
unlink(paste0(path_run,"*"), force = TRUE)

##- get the data
utils::download.file("https://github.com/gisma/gismaData/raw/master/uavRst/data/tutorial.zip",
                     paste0(path_run,"tutorial.zip"))
unzip(zipfile = paste0(path_run,"tutorial.zip"), exdir = path_run)


##- read chm data
chmR<- raster::raster(paste0(path_run,"chm_2.tif"))
tPos<- raster::raster(paste0(path_run,"treepos_2.tif"))

##- segmentation
crownsITC<- chmseg_ITC(chm = chmR,
                       EPSG =3064,
                       movingWin = 3,
                       TRESHSeed = 0.45,
                       TRESHCrown = 0.55,
                       minTreeAlt = 2,
                       maxCrownArea = 150)

##- visualisation
mapview::mapview(crownsITC,zcol="Height_m")
}
##+}
}
