% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_crownSegmentation.R
\name{chmSegmentationFU}
\alias{chmSegmentationFU}
\title{Create Tree Segmentation based on wrapping 'Fusion'}
\usage{
chmSegmentationFU(lasDir = NULL, cellsz = 0.5, perc = 37, fowsz = 3,
  fowfi = "mean", proj = "+init=epsg:25832", path = getwd(),
  Fp = "C:/FUSION/", extent = NULL)
}
\arguments{
\item{perc}{integer. n-th percentile of canopy heigth for heightbreak of Tree Segmentation (see Details)}

\item{fowsz}{integer. Cellsize for focal window (see Details)}

\item{fowfi}{function. The function fowfi should take multiple numbers and return a single number.
For example mean, modal, min or max}

\item{proj}{character proj4 string}

\item{path}{character. Set working directory [Example: path = "D:/TreeSeg/"]}

\item{Fp}{character. Directory for FUSION tools [Example: Fp = "D:/FUSION/"]}

\item{lasList}{filename (.las-file) or list of filenames (.txt-file)}

\item{cellcz}{integer. Cellsize for creating .dtm of las-point-cloud}
}
\description{
Method to create a Tree Segmentation using different FUSION tools. 
 The Tree Segmentation can be created from LAS-files and creates raster and shapefile output.
}
\details{
The "Create Tree Segmentation" (chmSegmentationFU) function works with a watershed segmentation algorithm (Vicent & Soille 1991). Basis for the calculation of "basins" is a canopy height model (CHM) and not point data files. From the input (.las-files) a canopy surface model using LIDAR point cloud ist created by the CanopyModel application from FUSION in .dtm format. The tool uses the returns with the highest elevation to compute the surface model. The applied /ground switch produces a canopy height model (CHM) by subtracting the earth surface model from the return elevation. A focal window with different arguments is used to calculate focal ("moving window") values for the neighborhood of focal cells. The Tree Segmentation is created by the chmSegmentationFU aplication from FUSION which used the filtered canopy heigth model and the heightbreak to calculate basins. The tool produces a metrics file in .csv-format, a shapefile containing high points and basin metrics, also another shapefile containing basin (crown) outline and a .tif raster map including basins.\cr\cr
\itemize{
\item{
 lasList: lasList has to be a .las-file or wildcard of .txt-file including different .las-files
 }
\item{
 cellsz: Grid size of input raster (CanopyHeigthModel) in Tree Segmantation (default=1) !!BE AWARE!!. Main parameter for chmSegmentationFU, adapt to tree species (size, crown size, etc.) necessary.
 }
\item{
 perc: default=37 (95th percentile), 38 (99th percentile), 35 (80th percentile), for more information see FUSION GridMatrix \url{https://w3.ual.es/GruposInv/ProyectoCostas/FUSION_manual.pdf}
 }
\item{
 fowsz: Focal uses a matrix of weights for the neighborhood of the focal cells. Fowsz ist the matrix of moving window. If fowsz is 3 a 3x3 window is used (default=3). For more details see: focal \url{https://www.rdocumentation.org/packages/raster/versions/2.6-7/topics/focal}
 }
\item{
 fowfi: Focal can use different function options like mean, modal, min or max for the calculation (default="mean")
 }
\item{
 proj: Set projection with EPSG-Code (default="epsg:25832" --> ETRS89 / UTM zone 32N) for more information see: \url{http://spatialreference.org/}
 }
\item{
 path: Creates a /output and /results folder. In /output all .tif-results and in /results are the .shp and basinmap.tif are saved.
 }
\item{
 Fp: Directory for FUSION tools. FUSION tools can be downloaded here: \url{http://forsys.sefs.uw.edu/fusion/fusionlatest.html}
 }
}
}
\examples{
{
 ## NOT RUN 
 ## i=list.files(path,pattern=".las$", full.names=FALSE)
 ## chmSegmentationFU(lasList=i,cellsz=c(1,3,5),perc=37,fowsz=3,fowfi="mean",proj="+init=epsg:25832",path,Fp)
}
}
\references{
{
 McGaughey (2006): FUSION/LDV: Software for LIDAR Data Analysis and Visualization.
 
 Vincent & Soille (1991): Watersheds in digital space: An efficient algorithm based on immersion simulations. IEEE Transactions on Pattern Analysis and Machine Intelligence, 13, 583-598. 
}
}
\seealso{
{
 focal \url{https://www.rdocumentation.org/packages/raster/versions/2.6-7/topics/focal} 
 
 FUSION \url{https://w3.ual.es/GruposInv/ProyectoCostas/FUSION_manual.pdf}
 
 EPSG \url{http://spatialreference.org/}
}
}
\author{
Santowski, A., Schupp, A. & C. Weber (2018), adapted to Linux and some other package issues Chris Reudenbach
}
