% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_colorspace.R
\name{colorspace}
\alias{colorspace}
\title{imagemagick based function to transform the RGB color space into another}
\usage{
colorspace(input = NULL, colorspace = c("CIELab", "CMY", "Gray", "HCL",
  "HSB", "HSI", "Log", "XYZ", "YUV"), compress = "None", verbose = FALSE,
  retRaster = TRUE)
}
\arguments{
\item{input}{character. name of a (Geo)Tiff containing RGB data channels}

\item{colorspace}{character. argument to determine colorspace see \href{https://www.systutorials.com/qa/1954/how-to-convert-tiff-images-from-rgb-color-cmyk-color-on-linux}{transform RGB colorspace} seealso \href{https://www.imagemagick.org/script/command-line-options.php#colorspace}{colorspace manual}}

\item{compress}{character. compression type depending on the imagemagick version the choices are: None, BZip, Fax, Group4, JPEG, JPEG2000, Lossless, LZW, RLE or Zip.}

\item{verbose}{be quiet}

\item{retRaster}{logical if true a rasterstack is returned}
}
\description{
Imagemagick is used for the calculation of the transformations. Please provide a tif file
}
\note{
you need to install imagemagick on your system, please look at see \href{https://www.systutorials.com/qa/1954/how-to-convert-tiff-images-from-rgb-color-cmyk-color-on-linux}{transform RGB colorspace} seealso \href{https://www.imagemagick.org/script/download.php}{imagemagick download section}.
}
\examples{
\dontrun{

# required packages
 require(uavRst)
 require(curl)
 
# project folder
projRootDir<-tempdir()

# create subfolders please mind that the pathes are exported as global variables
 paths<-link2GI::initProj(projRootDir = projRootDir,
                         projFolders = c("data/","data/ref/","output/","run/","las/"),
                         global = TRUE,
                         path_prefix = "path_")
## overide trailing backslash issue
 path_run<-ifelse(Sys.info()["sysname"]=="Windows", sub("/$", "",path_run),path_run)
 setwd(path_run)                                          
 unlink(paste0(path_run,"*"), force = TRUE
 
 url <- "https://github.com/gisma/gismaData/raw/master/uavRst/data/tutorial.zip"
 res <- curl::curl_download(url, paste0(path_run,"tutorial.zip"))
 unzip(zipfile = res, exdir = path_run)
 
# change colorspace from RGB to CIElab  
 r2 <- colorspace(input=paste0(path_run,"rgb_1.tif"),colorspace="CIELab")
 raster::plotRGB(r2)
}
}
