% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_colorspace.R
\name{colorspace}
\alias{colorspace}
\title{imagemagick based function to transform the RGB color space into another}
\usage{
colorspace(input = NULL, colorspace = c("CIELab", "CMY", "Gray", "HCL",
  "HSB", "HSI", "Log", "XYZ", "YUV"), compress = "None", verbose = FALSE,
  retRaster = TRUE)
}
\arguments{
\item{input}{character. name of a (Geo)Tiff containing RGB data channels}

\item{colorspace}{character. argument to determine colorspace see \href{https://www.systutorials.com/qa/1954/how-to-convert-tiff-images-from-rgb-color-cmyk-color-on-linux}{transform RGB colorspace} seealso \href{https://www.imagemagick.org/script/command-line-options.php#colorspace}{colorspace manual}}

\item{compress}{character. compression type depending on the imagemagick version the choices are: None, BZip, Fax, Group4, JPEG, JPEG2000, Lossless, LZW, RLE or Zip.}

\item{verbose}{be quiet}

\item{retRaster}{logical if true a rasterstack is returned}
}
\description{
Imagemagick is used for the calculation of the transformations. Please provide a tif file
}
\note{
you need to install imagemagick on your system, please look at see \href{https://www.systutorials.com/qa/1954/how-to-convert-tiff-images-from-rgb-color-cmyk-color-on-linux}{transform RGB colorspace} seealso \href{https://www.imagemagick.org/script/download.php}{imagemagick download section}.
}
\examples{
\dontrun{

# get temporary runtime folder
  setwd(tempdir())
  
# download some typical data as provided by the authority
 url<-"http://www.ldbv.bayern.de/file/zip/5619/DOP\%2040_CIR.zip"
 res <- curl::curl_download(url, "testdata.zip")
 unzip(res,junkpaths = TRUE,overwrite = TRUE)
 
# change colorspace from RGB to CIElab  
 colorspace(input="4490600_5321400.tif",colorspace="CIELab")
 
}
}
