% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_textures.R
\name{colorspace}
\alias{colorspace}
\title{imagemagick based function to transform the RGB color space into another}
\usage{
colorspace(input = NULL, colorspace = c("CIELab", "CMY", "Gray", "HCL",
  "HSB", "HSI", "Log", "XYZ", "YUV"), compress = "None", verbose = FALSE)
}
\arguments{
\item{input}{GeoTiff containing RGB data channels}

\item{colorspace}{argument to determine colorspace see \href{https://www.systutorials.com/qa/1954/how-to-convert-tiff-images-from-rgb-color-cmyk-color-on-linux}{transform RGB colorspace}}

\item{compress}{compression type}

\item{verbose}{be quiet}
}
\description{
imagemagick based function to transform the RGB color space into another
}
\note{
Imagemagick is used for the calculation of the transformations. Please provide a valid Tiff file
}
\examples{
\dontrun{
url<-"http://www.ldbv.bayern.de/file/zip/5619/DOP\%2040_CIR.zip"
res <- curl::curl_download(url, "testdata.zip")
unzip(res,junkpaths = TRUE,overwrite = TRUE)
colorspace(input=paste0(getwd(),"4490600_5321400.tif"),colorspace="CIELab")
}
}
