% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_bluedjinn.R
\name{correctLas}
\alias{correctLas}
\title{checks extension mismatch of a lasfile}
\usage{
correctLas(lasfiles,
  proj4 = "+proj=utm +zone=32 +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0")
}
\arguments{
\item{proj4}{correct proj4 string}

\item{lasfile}{filename of an las file}
}
\description{
checks extension mismatch of a lasfile
}
