% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_classify.R
\name{extractTrainData}
\alias{extractTrainData}
\title{extracts training data from a raster stack}
\usage{
extractTrainData(rasterStack = NULL, trainPlots = NULL, bnames = NULL,
  imgFN)
}
\arguments{
\item{rasterStack}{default is \code{NULL} rasterstack wcontaining all image data}

\item{trainPlots}{default is \code{NULL}  sp object providing training geometries}

\item{trainDataFn}{default is \code{file.path(tempdir(),"trainingDF.RData")} Name of the extracted training data file}
}
\description{
extracts training data from a raster stack.return extractTrainData returns a dataframe with all training data
}
\examples{
\dontrun{

trainingDF <- extractTrainData(rasterStack  = trainStack,
                               training     = training,
                               ids=c(1,2),
                               idLabel= c("green","nogreen")) 
}

}
\author{
Chris Reudenbach
}
