% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_basicTreeCrownFilter.R
\name{fa_basicTreeCrownFilter}
\alias{fa_basicTreeCrownFilter}
\title{applies minnimum height and min/max area thresholds on the raw tree crown geometries}
\usage{
fa_basicTreeCrownFilter(crownFn, minTreeAlt = 5, crownMinArea = 3,
  crownMaxArea = 150)
}
\arguments{
\item{crownFn}{filname of OGR comliant vector file}

\item{minTreeAlt}{minimum height in meter that will be regarded as tree}

\item{crownMinArea}{minimum area of crowns that is accepted}

\item{crownMaxArea}{maximum area of crowns that is accepted}

\item{funNames}{morphological parameters to be calculated}
}
\value{
fa_basicTreeCrownFilterbasically returns SPDF  with the crown polygons and all calculated parameters
}
\description{
applies minnimum height and min/max area thresholds on the raw tree crown geometries crowns
}
\author{
Chris Reudenbach
}
