% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_basicTreeCrownFilter.R
\name{fa_basicTreeCrownFilter}
\alias{fa_basicTreeCrownFilter}
\title{applies minnimum height and min/max area thresholds on the raw tree crown geometries}
\usage{
fa_basicTreeCrownFilter(crownFn, minTreeAlt = 10, crownMinArea = 5,
  crownMaxArea = 100, mintreeAltParam = "chmQ50", crownSTDW = NULL,
  opt = NULL, TAopt = NULL)
}
\arguments{
\item{crownFn}{filname of OGR comliant vector file}

\item{minTreeAlt}{minimum height in meter that will be regarded as tree}

\item{crownMinArea}{minimum area of crowns that is accepted}

\item{crownMaxArea}{maximum area of crowns that is accepted}

\item{mintreeAltParam}{parameter that is used for filtering mintreealt default ist Median "chmQ50"}

\item{crownSTDW}{parameter that optionally filters for the STDV of the crown altitudes default is NULL}

\item{opt}{threshold value for optional filter default is NULL}

\item{TAopt}{optional parameter that my be used for filtering default is NULL}
}
\value{
fa_basicTreeCrownFilterbasically returns SPDF  with the crown polygons and all calculated parameters
}
\description{
applies minnimum height and min/max area thresholds on the raw tree crown geometries crowns
}
\author{
Chris Reudenbach
}
