% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_crownMetrics.R
\name{fa_caMetrics}
\alias{fa_caMetrics}
\title{calculate morphometric features of polygons}
\usage{
fa_caMetrics(crownarea, funNames = c("length", "elongation",
  "eccentricityboundingbox", "solidity", "eccentricityeigen", "calliper",
  "rectangularity", "circularityharalick", "convexity"))
}
\arguments{
\item{crownarea}{sp polygon object}

\item{funNames}{("length","elongation","eccentricityboundingbox","solidity","eccentricityeigen","calliper","rectangularity","circularityharalick","convexity")}
}
