% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_classify.R
\name{get_counts}
\alias{get_counts}
\title{counts pixel values according to their classes}
\usage{
get_counts(ids = c(1, 2), position = NULL, imageFiles = NULL,
  buffersize = 1.5, outPrefix = "classified_index_", ext = ".tif",
  path = path_output, dropChars = 0)
}
\arguments{
\item{ids}{numeric. the ids used for the training}

\item{position}{sp. spatialpoint object containing the centre target positions}

\item{imageFiles}{raster* image/classification file}

\item{buffersize}{numeric. radius in meters around position}

\item{outPrefix}{character. out prefix string}

\item{ext}{character. extension}

\item{path}{character. output path}

\item{dropChars}{numeric. number of characters that should be dropped at the end of the filename}
}
\description{
counts pixel values according to their classes
}
\examples{
\dontrun{
# required packages
require(uavRst)
require(curl)
require(link2GI)

# project folde
projRootDir<-tempdir()

# create subfolders please mind that the pathes are exported as global variables
paths<-link2GI::initProj(projRootDir = projRootDir,
                         projFolders = c("data/","data/ref/","output/","run/","las/"),
                         global = TRUE,
                         path_prefix = "path_")
                         
# get the rgb image, chm and training data 
url1 <- "https://github.com/gisma/gismaData/raw/master/uavRst/data/tutorial_data.zip"
url2 <- "https://github.com/gisma/gismaData/raw/master/uavRst/data/position.zip"
res1 <- curl::curl_download(url1, paste0(path_run,"tutorial_data.zip"))
res2 <- curl::curl_download(url2, paste0(path_run,"position.zip"))
unzip(zipfile = res1,exdir = ifelse(Sys.info()["sysname"]=="Windows", sub("/$", "",path_run),path_run))
unzip(zipfile = res2,exdir = ifelse(Sys.info()["sysname"]=="Windows", sub("/$", "",path_run),path_run))
position <- raster::shapefile(paste0(path_run,"position.zip"))
imageFiles <-Sys.glob(paths = paste0(path_run,"rgb*","tif"))
imageFiles <- raster::stack(imageFiles)
df1<-get_counts(position = position,
              imageFiles = imageFiles,
              ext=".tif")
}
}
