% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_classify.R
\name{getCounts}
\alias{getCounts}
\title{trains model according to trainingdata and image stacks}
\usage{
getCounts(ids = c(1, 2), position = NULL, imageFiles = NULL,
  buffersize = 1.5, out_prefix = "classified_index_", ext = ".tif",
  path = path_output, dropChars = 0)
}
\arguments{
\item{ids}{ids}

\item{position}{position}

\item{imageFiles}{image files}

\item{buffersize}{size in meters around position}

\item{out_prefix}{out prefix string}

\item{ext}{extension}

\item{path}{output path}

\item{dropChars}{chars to drop}
}
\description{
trains model according to trainingdata and image stacks
}
\examples{
 
\dontrun{
df1<-getCounts(position = position,
              imageFiles = imageFiles,
              dropChars = 8,
              pre=pre,
              ext=".tif")
}
}
