% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_bluedjinn.R
\name{h_grass2tif}
\alias{h_grass2tif}
\title{converts GRASS raster to geotiff}
\usage{
h_grass2tif(runDir = NULL, layer = NULL, returnRaster = FALSE)
}
\arguments{
\item{runDir}{path of working directory}

\item{layer}{name GRASS raster}

\item{returnRaster}{return GRASS raster as an R raster object default = FALSE}
}
\description{
converts GRASS raster to geotiff
}
