% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_karim.R
\name{h_grass2tif}
\alias{h_grass2tif}
\alias{h_grass2tif}
\title{converts GRASS raster to geotiff}
\usage{
h_grass2tif(runDir = NULL, layer = NULL)

h_grass2tif(runDir = NULL, layer = NULL)
}
\arguments{
\item{runDir}{path of working directory}

\item{layer}{name GRASS raster}

\item{returnRaster}{return GRASS raster as an R raster object default = FALSE}

\item{runDir}{path of working directory}

\item{layer}{name GRASS raster}
}
\description{
converts GRASS raster to geotiff

converts geotiff to GRASS raster
}
