% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_karim.R
\name{linkBuilder}
\alias{linkBuilder}
\title{creates all link2GI links}
\usage{
linkBuilder(links = NULL, simple = TRUE, linkItems = c("saga", "grass7",
  "otb", "gdal"), sagaArgs = "default", grassArgs = "default",
  otbArgs = "default",
  gdalArgs = "quiet = TRUE, \\n                                     returnPaths = TRUE")
}
\arguments{
\item{links}{character links}

\item{linkItems}{character list of c("saga","grass7","otb","gdal")
@param simple boolean true  make all
@param sagaArgs sagaArgs
@param grassArgs grassArgs}

\item{otbArgs}{otbArgs}

\item{gdalArgs}{gdalArgs}
}
\description{
converts SAGA raster to R raster object
}
