% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_bluedjinn.R
\name{make_bandnames}
\alias{make_bandnames}
\title{create name vector corresponding to the training image stack
create vector containing the names of the image stack claculated using \code{\link{calc_ext}}}
\usage{
make_bandnames(
  rgbi = NA,
  bandNames = NA,
  stat = FALSE,
  morpho = NA,
  edge = NA,
  rgbTrans = NA,
  dem = NA,
  l_raster = NA
)
}
\arguments{
\item{rgbi}{character. codes of the RGB indices}

\item{bandNames}{character.  band names}

\item{stat}{character.  stat codes}

\item{morpho}{character.  morpho codes}

\item{edge}{character.  edge codes}

\item{rgbTrans}{character.  rgbTrans codes}

\item{dem}{charater. dem codes}

\item{l_raster}{number of raster layer that exist}
}
\description{
create name vector corresponding to the training image stack
create vector containing the names of the image stack claculated using \code{\link{calc_ext}}
}
\keyword{internal}
