% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_classify.R
\name{makebNames}
\alias{makebNames}
\title{create name vector corresponding to the training image stack}
\usage{
makebNames(rgbi = NA, bandNames = NA, stat = FALSE, morpho = NA,
  edge = NA, RGBtrans = NA)
}
\arguments{
\item{rgbi}{default is  NA}

\item{bandNames}{band names}

\item{stat}{band names}

\item{morpho}{band names}

\item{edge}{band names}

\item{RGBtrans}{RGBtrans}
}
\description{
create name vector corresponding to the training image stack
}
