% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_pc2DSM.R
\name{pc2dsm}
\alias{pc2dsm}
\title{Create a Digital Surface Model from a UAV generated point cloud}
\usage{
pc2dsm(lasDir = NULL, gisdbase_path = NULL, GRASSlocation = "tmp/",
  projSubFolder = c("data/", "output/", "run/", "las/"), grid_size = "0.25",
  grass_lidar_method = "mean", grass_lidar_pth = 90,
  saga_spline_level_max = "4", otb_gauss_radius = "0.5",
  type_smooth = "gauss", dsm_minalt = 0, dsm_maxalt = 4000,
  dsm_area = FALSE, cutExtent = NULL,
  proj4 = "+proj=utm +zone=32 +datum=WGS84 +units=m +no_defs",
  gisdbase_exist = FALSE, path_lastools = NULL, giLinks = NULL,
  verbose = FALSE)
}
\arguments{
\item{lasDir}{default is \code{NULL} path  to the laz/las file(s)}

\item{gisdbase_path}{gisdbase will be linked or created depending on \code{gisdbase_exist}}

\item{GRASSlocation}{location will be linked or created depending on \code{gisdbase_exist}}

\item{projSubFolder}{subfolders that will be created/linked for R related GRASS processing}

\item{grid_size}{resolution for raster operations}

\item{grass_lidar_method}{statistical method to sample the Lidar data}

\item{grass_lidar_pth}{grass lidar aggregation column percentile}

\item{saga_spline_level_max}{default is 9 number ob spline iterations}

\item{otb_gauss_radius}{default is \code{0.5} radius of otb smoothing filter in meter}

\item{type_smooth}{default is \code{gauss} alternatives are \code{spline} or for no smoothing at all \code{no}}

\item{dsm_minalt}{default is \code{0}, minimum DTM altitude accepted}

\item{dsm_maxalt}{dsm maximum altitude}

\item{dsm_area}{default \code{FALSE} generate polygon of valid DSM data}

\item{cutExtent}{clip area}

\item{proj4}{default is EPSG \code{32632} any valid proj4 string that is assumingly the correct one}

\item{gisdbase_exist}{default is  \code{FALSE} switch if gisdbase is created or  linked only}

\item{path_lastools}{character folder containing the Windows binary files of the LAStools}

\item{giLinks}{list of GI tools cli pathes as generated by a full call of link2GI}

\item{verbose}{to be quiet (1)}
}
\description{
Create a Digital Surface Model from a UAV generated point cloud. Basically returns a  DSM. 
It uses the 'r.in.lidar' function to calculate LiDAR derived raster grids. 
It creates a \code{raster*} object.
}
\examples{
\dontrun{
# create a DSM based on a uav point cloud 
 pc2dsm(lasDir =  lasDir,
        gisdbase_path = "~/temp6/GRASS7",
        GRASSlocation = "tmp/",
        projSubFolder = c("data/","output/","run/","las/"),
        grid_size = "0.05",
        gisdbase_exist = FALSE)
}

}
\seealso{
\href{https://grass.osgeo.org/grass70/manuals/r.in.lidar.html}{r.in.lidar help}
}
\author{
Chris Reudenbach
}
