% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foa_pc2DSM.R
\name{pc2DSM}
\alias{pc2DSM}
\title{Create a Digital Surface Model from a UAV generated point cloud}
\usage{
pc2DSM(lasDir = NULL, gisdbase_path = NULL, GRASSlocation = "tmp/",
  projFolder = c("data/", "output/", "run/", "las/"), grid_size = "0.5",
  saga_spline_level_max = "9", otb_gauss_radius = "0.5",
  type_smooth = "otb_gauss", dsm_minalt = 0, dsm_maxalt = 4000,
  dsm_area = FALSE,
  proj4 = "+proj=utm +zone=32 +datum=WGS84 +units=m +no_defs",
  gisdbase_exist = FALSE, path_lastools = NULL)
}
\arguments{
\item{lasDir}{default is \code{NULL} path  to the laz/las file(s)}

\item{gisdbase_path}{gisdbase will be linked or created depending on \code{gisdbase_exist}}

\item{GRASSlocation}{location will be linked or created depending on \code{gisdbase_exist}}

\item{projFolder}{subfolders in gisdbase for R related processing}

\item{grid_size}{resolution for raster operations}

\item{saga_spline_level_max}{default is 9 number ob spline iterations}

\item{otb_gauss_radius}{default is \code{0.5} radius of otb smoothing filter in meter}

\item{type_smooth}{default is \code{otb_gauss} alternatives is \code{saga_spline}}

\item{dsm_minalt}{default is \code{0}, minimum DTM altitude accepted}

\item{dsm_area}{default \code{FALSE} generate polygon of valid DSM data}

\item{proj4}{default is EPSG \code{32632} any valid proj4 string that is assumingly the correct one}

\item{gisdbase_exist}{default is  \code{FALSE} switch if gisdbase is created or  linked only}

\item{dsm_minalt}{default is \code{4000}, maximum DTM altitude accepted}
}
\value{
pc2DSM basically returns a  DSM
}
\description{
Create a Digital Surface Model from a UAV generated point cloud
}
\examples{
\dontrun{
# create a DSM based on a uav point cloud 
 pc2DSM(lasDir =  lasDir,
        gisdbase_path = "~/temp6/GRASS7",
        GRASSlocation = "tmp/",
        projFolder = c("data/","output/","run/","las/"),
        grid_size = "0.05",
        gisdbase_exist = FALSE)
}

}
\author{
Chris Reudenbach
}
