% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDTM.R
\name{pc2DTM}
\alias{pc2DTM}
\title{calcualte and post-classifies the morphological structure of raw tree crowns}
\usage{
pc2DTM(runDir,currentP,allP)
}
\arguments{
\item{level_max}{number ob spline iterations}

\item{gridsize}{resolution of the DTM raster}

\item{path_lastools}{directory for the windows lastools}

\item{cores}{number of cores that will be used}
}
\value{
pc2DTM basically returns a DEM and DSM
}
\description{
calcualte and post-classifies the morphological structure of raw tree crowns
}
\author{
Chris Reudenbach
}
