% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_classify.R
\name{predict_rgb}
\alias{predict_rgb}
\title{classify images using raster predict}
\usage{
predict_rgb(imageFiles = NULL, model = NULL, inPrefix = "index_",
  outPrefix = "classified_", bandNames = NULL)
}
\arguments{
\item{imageFiles}{raster*. imagestack for classification purposes must contain the required bands as needed by the model.}

\item{model}{model. classification model}

\item{inPrefix}{character. in frefix  string}

\item{outPrefix}{character. out prefix string}

\item{bandNames}{character. band names}
}
\description{
classify images using raster predict
}
\examples{
\dontrun{
# required packages
require(uavRst)
require(curl)
require(link2GI)

# project folder
projRootDir<-tempdir()

# create subfolders pls notice the pathes are exported as global variables
paths<-link2GI::initProj(projRootDir = projRootDir,
                         projFolders = c("data/","data/ref/","output/","run/","las/"),
                         global = TRUE,
                         path_prefix = "path_")
# overide trailing backslash issue
 path_run<-ifelse(Sys.info()["sysname"]=="Windows", sub("/$", "",path_run),path_run)
 setwd(path_run)                                                 
 unlink(paste0(path_run,"*"), force = TRUE)
 
# get the data rgb image, chm data and training data 
 url <- "https://github.com/gisma/gismaData/raw/master/uavRst/data/ffs.zip"
 res <- curl::curl_download(url, paste0(path_run,"ffs.zip"))
 unzip(zipfile = res, exdir = path_run)
 

imageFile <- paste0(path_run,"predict.tif")
load(paste0(path_run,"tutorialbandNames.RData"))
tutorialModel<-readRDS(file = paste0(path_run,"tutorialmodel.rds"))

## start prediction taking the non optimized model
predict_rgb(imageFiles=imageFile,
            model = tutorialModel[[1]],
            bandNames = bandNames)    
}                                         
}
