% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_classify.R
\name{predict_rgb}
\alias{predict_rgb}
\title{classify images using raster predict}
\usage{
predict_rgb(imageFiles = NULL, model = NULL, inPrefix = "index_",
  outPrefix = "classified_", bandNames = NULL)
}
\arguments{
\item{imageFiles}{raster*. imagestack for classification purposes must contain the required bands as needed by the model.}

\item{model}{model. classification model}

\item{inPrefix}{character. in frefix  string}

\item{outPrefix}{character. out prefix string}

\item{bandNames}{character. band names}
}
\description{
classify images using raster predict
}
\examples{
\dontrun{

# required packages
require(uavRst)
require(curl)
require(link2GI)

# project folder
projRootDir<-tempdir()

# create subfolders pls notice the pathes are exported as global variables
paths<-link2GI::initProj(projRootDir = projRootDir,
                         projFolders = c("data/","data/ref/","output/","run/","las/"),
                         global = TRUE,
                         path_prefix = "path_")

# get the data rgb image, chm data and training data 

url <- "https://github.com/gisma/gismaData/raw/master/uavRst/data/tutorial_predict.zip"
res <- curl::curl_download(url, paste0(path_run,"predict.zip"))
unzip(zipfile = res,exdir = paste0(path_run,"predict.zip"))

# create the links to the GI software
giLinks<-uavRst::get_gi()

# get the files  
 imageTrainStack <- list()
 geomTrainStack <- list()
 imageTrainFiles <- list.files(pattern="[.]envi$", path=path_run, full.names=TRUE)
 geomTrainFiles <- list.files(pattern="[.]shp$", path=path_run, full.names=TRUE)
 load(paste0(path_run,"rgbImgbandNames.RData"))
 load(file = paste0(path_output,prefixRun ,"_model_final",".RData"))
 
 # start prediction
 predict_rgb(imageFiles=imageFiles,
             model = model_final,
             inPrefix = "index_",
             outPrefix = "classified_",
             bandNames = bandNames)
                                                 
}
}
