% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_bluedjinn.R
\name{r2saga}
\alias{r2saga}
\title{converts SAGA raster to R raster object}
\usage{
r2saga(x, fn)
}
\arguments{
\item{x}{raster object}

\item{fn}{filname without extension}
}
\description{
converts SAGA raster to R raster object
}
