% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_bluedjinn.R
\name{r2saga}
\alias{r2saga}
\title{converts R raster* objects to SAGA raster}
\usage{
r2saga(x, fn, path_run = tempdir())
}
\arguments{
\item{x}{raster object}

\item{fn}{filname name without extension}

\item{path_run}{character. path used for runtime operations}
}
\description{
converts R raster* objects to SAGA raster
}
\keyword{internal}
