% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_karim.R
\name{SAGA2R}
\alias{SAGA2R}
\title{converts SAGA raster to R raster object}
\usage{
SAGA2R(fn, ext)
}
\arguments{
\item{fn}{filname without extension}

\item{ext}{extent of the raster in R notation}
}
\description{
converts SAGA raster to R raster object
}
