% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_bluedjinn.R
\name{saga2r}
\alias{saga2r}
\title{converts SAGA raster to R raster object}
\usage{
saga2r(fn, ext, path_run = tempdir(), gdalLinks = NULL)
}
\arguments{
\item{fn}{filname without extension}

\item{ext}{extent of the raster in R notation}

\item{path_run}{character. path used for runtime operations}

\item{gdalLinks}{list of gdal installations}
}
\value{
returned saga raster file as an raster* object
}
\description{
converts SAGA raster to R raster object
}
\keyword{internal}
