% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_leafSel.R
\name{select}
\alias{select}
\title{selecting vector features}
\usage{
select(mapCenter = c(50.80801, 8.72993), zoom = 12, position = "topright",
  maplayer = c("CartoDB.Positron", "OpenStreetMap", "Esri.WorldImagery",
  "Thunderforest.Landscape", "OpenTopoMap"), overlay = NULL, cex = 10,
  alpha = 0.6, opacity = 0.7, color = "red")
}
\arguments{
\item{zoom}{initial zoom level}

\item{position}{place to put the toolbar (topright, topleft, bottomright, bottomleft)}

\item{maplayer}{string as provided by leaflet-provider}

\item{overlay}{optional sp object}

\item{mapcenter}{c(lat,lon) central point of the leaflet map}

\item{intersection}{enable/disable th possibility to overlay lines or polygons}
}
\description{
select provides a simple interatcive way to select and deselect features from a vector data set. the selection is exportes to a Json format and also written in a textpox for copy and paste issues
}
\examples{

# get meuse data
data(meuse)
coordinates(meuse) <- ~x+y
proj4string(meuse) <- CRS("+init=epsg:28992")


# preset for digitizing uav flight areas

#' # preset for digitizing extents
select(preset="ext")

}
\author{
Chris Reudenbach
}
