% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_leafSel.R
\name{select}
\alias{select}
\title{selecting vector features}
\usage{
select(mapCenter = c(50.80801, 8.72993), zoom = 15, line = TRUE,
  rectangle = FALSE, poly = FALSE, circle = FALSE, point = FALSE,
  remove = TRUE, hidemenu = TRUE, position = "topright",
  maplayer = c("CartoDB.Positron", "OpenStreetMap", "Esri.WorldImagery",
  "Thunderforest.Landscape", "OpenTopoMap"), overlay = NULL,
  features = NULL, cex = 4, lwd = 1, alpha = 0.6, opacity = 0.7,
  color = "blue")
}
\arguments{
\item{zoom}{initial zoom level}

\item{line}{enable/disable the draw tool line tool}

\item{poly}{enable/disable the draw polygon tool}

\item{circle}{enable/disable the draw circle tool}

\item{point}{enable/disable the draw point tool}

\item{remove}{enable/disable the remove feature of the draw tool}

\item{position}{place to put the toolbar (topright, topleft, bottomright, bottomleft)}

\item{maplayer}{leaflet-provider maplayers}

\item{overlay}{sp object  for selection}

\item{cex}{size of features}

\item{alpha}{0.6,}

\item{opacity}{0.7,}

\item{color}{"blue"}

\item{mapcenter}{c(lat,lon) central point of the leaflet map}

\item{intersection}{enable/disable th possibility to overlay lines or polygons}
}
\description{
select provides a simple interatcive way to select and deselect features from a vector data set. the selection is exportes to a Json format and also written in a textpox for copy and paste issues
}
\examples{

# get meuse data
data(meuse)
coordinates(meuse) <- ~x+y
proj4string(meuse) <- CRS("+init=epsg:28992")


# select some points
 select(overlay = meuse, zoom = 13,position = "bottomright", cex = 2, color = "green")
 
 # re-import selection saved as "test.json"
 import <-rgdal::readOGR(dsn = path.expand("~/test.json"), layer = "OGRGeoJSON")
 
 # select some polygons
 uavRst::select(overlay = gadmCHE,zoom = 7)
 
 #'  # re-import selection saved as "ch_sel.json"
 import <-rgdal::readOGR(dsn = path.expand("~/ch_sel.json"), layer = "OGRGeoJSON")

}
\author{
Chris Reudenbach
}
