% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_bluedjinn.R
\name{shape2grass}
\alias{shape2grass}
\title{converts OGR to GRASS vector}
\usage{
shape2grass(runDir = NULL, layerName = NULL)
}
\arguments{
\item{runDir}{path of working directory}

\item{layeName}{name GRASS raster}
}
\description{
converts OGR to GRASS vector
}
\keyword{internal}
