% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_bluedjinn.R
\name{sp_line}
\alias{sp_line}
\title{Create an spatiallineobject from 2 points}
\usage{
sp_line(startPoint, endPoint, ID,
  proj4 = "+proj=longlat +datum=WGS84 +no_defs", export = FALSE)
}
\arguments{
\item{startPoint}{coordinate of first point (c(50.1,8.1))}

\item{endPoint}{vector. coordinate of second point c(50.2,8.0)}

\item{ID}{id of line}

\item{proj4}{proj4 string}

\item{export}{write shafefile default = F}
}
\description{
create an spatiallineobject from 2 points, optional export as shapefile
}
\author{
Chris Reudenbach
}
\keyword{internal}
