% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_karim.R
\name{sp_point}
\alias{sp_point}
\title{create an spatialpointobject from 1 points}
\usage{
sp_point(lon, lat, ID = "point",
  proj4 = "+proj=longlat +datum=WGS84 +no_defs", export = FALSE)
}
\arguments{
\item{ID}{name of point}

\item{export}{write shafefile default = F}

\item{p1}{coordinate of first point}
}
\description{
create an spatialpointobject from 1 points, optional export as shapefile
}
