% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_spc2dtm.R
\name{spc2dtm}
\alias{spc2dtm}
\title{Create a Digital Terrain Model from UAV generated point clouds by minimum sampling}
\usage{
spc2dtm(laspcFile = NULL, gisdbase_path = NULL, tension = 20,
  cutExtent = NULL, sampleGridSize = 25, targetGridSize = 0.5,
  projFolder = NULL,
  proj4 = "+proj=utm +zone=32 +datum=WGS84 +units=m +no_defs",
  giLinks = NULL, verbose = FALSE)
}
\arguments{
\item{laspcFile}{default is \code{NULL} path  to the laz/las file(s)}

\item{gisdbase_path}{default is \code{NULL} root directory of the project. NOTE the function creates two subfolder named \code{run} and \code{output}}

\item{tension}{numeric. tension of spline interpolation.}

\item{cutExtent}{clip area}

\item{sampleGridSize}{resolution extraction raster}

\item{targetGridSize}{the resolution of the target DTM raster}

\item{projFolder}{subfolders that will be created/linked for R related GRASS processing}

\item{proj4}{default is EPSG 32632, any valid proj4 string that is assumingly the correct one}

\item{giLinks}{list of GI tools cli pathes, default is NULL}

\item{verbose}{to be quiet (1)}
}
\description{
Create a Digital Terrain Model from a high density point cloud as typically derived by an optical UAV retrieval. Due to the poor estimation of ground points 
a minimum samopling approach is applied. It retrieves on a coarse sampling gridsize the minimum value and interpolates on these samples a surface grid with a higher target 
resolution. this is a kind of an try and error process and provides fairly good results if the point cloud shows at least some real surface points on a not to coarse grid.
}
\examples{
\dontrun{
spc2dtm(laspcFile =  "~/path/to/lasdata",
       gisdbase_path = "~/temp5",
       thin_with_grid = "0.5",
       level_max = "5" ,
       grid_size = "0.5")
}
}
\author{
Chris Reudenbach
}
