% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_lidr_makros.R
\name{tile_eq}
\alias{tile_eq}
\title{divide a lidR las* or lascatalog object by given factors}
\usage{
tile_eq(
  cgs,
  factor_x = 4,
  factor_y = 4,
  outpath = NULL,
  las_fn = "cut",
  proj4 = "+init=epsg:32632"
)
}
\arguments{
\item{cgs}{filename of an las file}

\item{factor_x}{factor to divide in x direction}

\item{factor_y}{factor to divide in y direction}

\item{outpath}{path to write to if null nothing is written}

\item{las_fn}{filname of lasfile if written}

\item{proj4}{correct proj4 string}
}
\description{
divide a lidR las* object by given factors. The resulting tiles are  written to the outpath folder
}
\examples{
\dontrun{
cut_eq(cgs = lidR::readLAS(las_file),factor_x = 4,factor_y = 4 )
}
}
