% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainModel.R
\name{trainModel}
\alias{trainModel}
\title{trains model according to trainingdata and image stacks}
\usage{
trainModel(trainingDF = NULL, predictors = c("R", "G", "B"),
  response = "ID", spaceVar = "FN", names = c("ID", "R", "G", "B", "A",
  "FN"), noLoc = 3, cl_method = "rf", metric_ffs = "kappa",
  metric_caret = "ROC", pVal = 0.5, prefin = "final_", preffs = "ffs_",
  modelSaveName = "model.RData")
}
\arguments{
\item{predictors}{default is \code{c("R","G","B")}}

\item{response}{default is \code{"ID"}}

\item{spaceVar}{default is \code{"FN"}}

\item{names}{default is \code{c("ID","R","G","B","A","FN")}}

\item{noLoc}{default is \code{3}}

\item{cl_method}{default is \code{"rf"}}

\item{metric_ffs}{default is \code{"kappa"}}

\item{metric_caret}{default is \code{ "ROC"}}

\item{pVal}{default is \code{ 0.5}}

\item{prefin}{default is \code{"final_"}}

\item{preffs}{default is \code{"ffs_"}}

\item{modelSaveName}{default is \code{"model.RData" }}

\item{imageFiles}{imagestacke for classification}

\item{model}{classification model}

\item{in_prefix}{in frefix  string}

\item{out_prefix}{out prefix string}

\item{bandNames}{band names 
trainingDF =NULL,}
}
\description{
trains model according to trainingdata and image stacks
}
\examples{
 
result<-  trainModel(trainingDF =trainingDF,
                     predictors   = c("R","G","B","VARI","NDTI","TGI","GLI","NGRDI","GLAI","Mean","Variance","Skewness","Kurtosis"),
                     response     = "ID",
                     spaceVar     = "FN",
                     names = c("ID","R","G","B","A","VARI","NDTI","TGI","GLI","NGRDI","GLAI","Mean","Variance","Skewness","Kurtosis","FN"),
                     noLoc        = length(imageTrainFiles),
                     cl_method    = "rf",
                     metric_ffs   = "kappa",
                     metric_caret = "ROC",
                     pVal         = 0.5) 
}
