% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_classify.R
\name{trainModel}
\alias{trainModel}
\title{forward feature selection based random forest model training}
\usage{
trainModel(trainingDF = NULL, predictors = c("R", "G", "B"),
  response = "ID", spaceVar = "FN", names = c("ID", "R", "G", "B", "A",
  "FN"), noLoc = 3, sumFunction = "twoClassSummary", pVal = 0.5,
  prefin = "final_", preffs = "ffs_", modelSaveName = "model.RData",
  runtest = FALSE, noClu = 3)
}
\arguments{
\item{trainingDF}{dataframe containing training data}

\item{predictors}{vector of predictor names as given by the header of the training data table}

\item{response}{name of response variable as given by the header of the training data table}

\item{spaceVar}{name of the spcetime splitting vatiable as given by the header of the training data table}

\item{names}{all names of the dataframe header}

\item{noLoc}{number of locations to leave out usually nuber of dicrete trainings locations/images}

\item{pVal}{used part of the training data  default is \code{ 0.5}}

\item{prefin}{name pattern used for model default is \code{"final_"}}

\item{preffs}{name pattern used for ffs default is \code{"ffs_"}}

\item{modelSaveName}{name pattern used for saving the model default is \code{"model.RData" }}

\item{runtest}{logical default is false, if set a external validation will be performed}

\item{metric}{accuracy metrics for ffs for classification  default is \code{"kappa"}}

\item{nrclu}{number of cluster to be used}
}
\description{
trainModel is a wrapper function for a simple use of the forwatrd feature sselection approach 
of training random forest classification models. This validation is particulary suitable for
leave-location-out cross validations where variable selection
MUST be based on the performance of the model on the hold out station.
See \href{https://doi.org/10.1016/j.envsoft.2017.12.001}{Meyer et al. (2018)}
for further details.
This is in fact the case while using time space variable vegetation patterns for classification purposes. 
For the uav based RGB/NIR imagery it provides an optimized preconfiguration for the classification goals.
}
\examples{
 
#' \dontrun{
result<-  trainModel(trainingDF =trainingDF,
                     predictors   = c("R","G","B"),
                     response     = "ID",
                     spaceVar     = "FN",
                     names        = c("ID","R","G","B","A","FN"), 
                     noLoc        = length(imageTrainFiles),
                     metric_ffs   = "kappa",
                     pVal         = 0.5) 
                 }
}
