% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uavRst-package.R
\docType{package}
\name{uavRst}
\alias{uavRst}
\alias{uavRst-package}
\title{Unmanned Aerial Vehicle Remote Sensing Tools - awesome tools to manipulate and analyze RGB imagery and point clouds}
\description{
The package provides some R-GI functionality for dealing with data as retrieved by Unmanned Aerial Vehicle VIS/NIR Remote Sensing
}
\details{
Unmanned Aerial Vehicle Remote Sensing Tools
}
\author{
Hanna Meyer, Lars Opgenoorth, Chris Reudenbach, Environmental Informatics Marburg \cr
\cr
\emph{Maintainer:} Chris Reudenbach \email{reudenbach@uni-marburg.de}
}
\keyword{package}
