% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_bluedjinn.R
\name{xyz2tif}
\alias{xyz2tif}
\title{Read and Convert xyz DEM/DSM Data as typically provided by the Authorities}
\usage{
xyz2tif(xyzFN = NULL, epsgCode = "25832")
}
\arguments{
\item{xyzFN}{ASCII tect file with xyz values}

\item{epsgCode}{"25832"}
}
\description{
Read xyz data and generate a raster  \code{Raster*} object.
}
\examples{
\dontrun{
#get some typical data as provided by the authority
setwd(tempdir())
url<-"http://www.ldbv.bayern.de/file/zip/10430/DGM_1_ascii.zip"
res <- curl::curl_download(url, "testdata.zip")
file<- unzip(res,list = TRUE)$Name[2]
unzip(res,files = file,  overwrite = TRUE)
head(read.csv(file))
xyz2tif(file,epsgCode = "31468")
}
}
